Block Design for Response-Adaptive Randomization for Binomial Data

Simulation for binomial counts for block design for response-adaptive randomization with time as a confounding

binomialRAR(p_control, p_treatment, N_total, block_number = 4,
  drift = 0, simulation = 10000, zvalue = c(1, 1.5, 2, 2.5),
  rand_ratio = c(1, 1.5, 2, 2.5), conf_int = 0.95,
  alternative = "less", replace = FALSE)

Arguments

p_control
scalar. Proportion of events under the control arm.

p_treatment
scalar. Proportion of events under the treatment arm.

N_total
scalar. Total sample size.

block_number
scalar. Number of blocks or time levels. The default is set to 4. If block_number is set to 1. This is a traditional RCT design.

drift
scalar. The increase or decrease in proportion of event over time. In this case, the proportion of failure changes in each block by the number of patient accured over the totoal sample size. The full drift effect is seen in the final block.

simulation
scalar. Number of simulation to be ran. The default is set to 10000.

zvalue
vector. The z-value cutoff for corrected chi-square test statistics.

rand_ratio
vector. The randomization ratio is set based on the corrected chi-square test statistic. The length of rand_ratio should be the same length of zvalue.

conf_int
scalar. Confidence level of the interval.

alternative
character. A string specifying the alternative hypothesis, must be one of "less" (default) or "greater".

replace
character. should sampling be with replacement? If replace is set to FALSE (default), the 0 for control, 1 for treatment is replicated to the closest integer and this vector is sampled with no replacement. If replace is set to TRUE, the sampling is done based on randomization ratio provided with replacement.

Value

a list with details on the simulation.

power

scalar. The power of the trial, ie. the proportion of success over the number of simulation ran.

p_control_estimate

scalar. The estimated proportion of events under the control group.

p_treatment_estimate

scalar. The estimated proportion of events under the treatment group.

N_enrolled

vector. The number of patients enrolled in the trial (sum of control and experimental group for each simulation. )

N_control

vector. The number of patients enrolled in the control group for each simulation.

N_control

vector. The number of patients enrolled in the experimental group for each simulation.

Examples

binomialRAR(p_control = 0.7, p_treatment = 0.65, N_total = 200,
            block_number = 2, simulation = 100)

#> $power
#> [1] 0.12
#> 
#> $p_control_estimate
#>   [1] 0.6000000 0.7600000 0.7000000 0.6900000 0.7974684 0.6600000 0.7600000
#>   [8] 0.7200000 0.7200000 0.6700000 0.6000000 0.7000000 0.7700000 0.7341772
#>  [15] 0.7333333 0.7215190 0.7721519 0.8400000 0.6900000 0.6400000 0.6700000
#>  [22] 0.6500000 0.7600000 0.6600000 0.7200000 0.7500000 0.6700000 0.7000000
#>  [29] 0.6636364 0.7222222 0.6600000 0.6500000 0.6500000 0.7100000 0.6700000
#>  [36] 0.7900000 0.6600000 0.7300000 0.7400000 0.7111111 0.6900000 0.8072289
#>  [43] 0.7000000 0.6300000 0.7100000 0.7000000 0.7200000 0.6818182 0.6300000
#>  [50] 0.6500000 0.7000000 0.6300000 0.6300000 0.7555556 0.7777778 0.7848101
#>  [57] 0.6800000 0.6545455 0.7100000 0.7555556 0.6909091 0.6800000 0.7500000
#>  [64] 0.6900000 0.7200000 0.7888889 0.7444444 0.7444444 0.6818182 0.8000000
#>  [71] 0.7222222 0.7300000 0.8200000 0.6600000 0.6600000 0.6700000 0.7400000
#>  [78] 0.6700000 0.6800000 0.6300000 0.6900000 0.6500000 0.7333333 0.6700000
#>  [85] 0.7600000 0.7100000 0.7600000 0.7000000 0.7100000 0.5900000 0.7700000
#>  [92] 0.7600000 0.6700000 0.7500000 0.6700000 0.7500000 0.6900000 0.7000000
#>  [99] 0.6900000 0.6500000
#> 
#> $p_treatment_estimate
#>   [1] 0.5600000 0.6500000 0.6500000 0.6200000 0.6198347 0.7000000 0.6600000
#>   [8] 0.6500000 0.7000000 0.6500000 0.5900000 0.5800000 0.6000000 0.6363636
#>  [15] 0.5727273 0.6280992 0.6611570 0.5200000 0.6700000 0.6200000 0.6200000
#>  [22] 0.6900000 0.7300000 0.5700000 0.6700000 0.6800000 0.6000000 0.6600000
#>  [29] 0.7333333 0.6272727 0.6500000 0.6300000 0.6500000 0.6600000 0.6800000
#>  [36] 0.6800000 0.6800000 0.6500000 0.6400000 0.6454545 0.6400000 0.5982906
#>  [43] 0.6700000 0.6400000 0.6600000 0.6400000 0.6900000 0.7333333 0.6300000
#>  [50] 0.6300000 0.6600000 0.6200000 0.6400000 0.6909091 0.6909091 0.5289256
#>  [57] 0.6300000 0.6777778 0.6200000 0.6636364 0.6666667 0.6500000 0.6900000
#>  [64] 0.7200000 0.7000000 0.6272727 0.6727273 0.6363636 0.7777778 0.7000000
#>  [71] 0.5636364 0.6900000 0.5400000 0.6700000 0.6300000 0.6600000 0.6500000
#>  [78] 0.7000000 0.6300000 0.6800000 0.6100000 0.6700000 0.7000000 0.7200000
#>  [85] 0.7000000 0.6700000 0.6600000 0.6100000 0.6000000 0.5800000 0.6800000
#>  [92] 0.6200000 0.6600000 0.6400000 0.6200000 0.7000000 0.6800000 0.6500000
#>  [99] 0.6800000 0.6000000
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 100
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 100 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1] 100 100 100 100  79 100 100 100 100 100 100 100 100  79  90  79  79  50
#>  [19] 100 100 100 100 100 100 100 100 100 100 110  90 100 100 100 100 100 100
#>  [37] 100 100 100  90 100  83 100 100 100 100 100 110 100 100 100 100 100  90
#>  [55]  90  79 100 110 100  90 110 100 100 100 100  90  90  90 110 100  90 100
#>  [73]  50 100 100 100 100 100 100 100 100 100  90 100 100 100 100 100 100 100
#>  [91] 100 100 100 100 100 100 100 100 100 100
#> 
#> $N_treatment
#>   [1] 100 100 100 100 121 100 100 100 100 100 100 100 100 121 110 121 121  50
#>  [19] 100 100 100 100 100 100 100 100 100 100  90 110 100 100 100 100 100 100
#>  [37] 100 100 100 110 100 117 100 100 100 100 100  90 100 100 100 100 100 110
#>  [55] 110 121 100  90 100 110  90 100 100 100 100 110 110 110  90 100 110 100
#>  [73]  50 100 100 100 100 100 100 100 100 100 110 100 100 100 100 100 100 100
#>  [91] 100 100 100 100 100 100 100 100 100 100
#>