Block Design for Response-Adaptive Randomization for Binomial Data

Simulation for binomial counts for block design for response-adaptive randomization with time as a confounding

binomialfreq(p_control, p_treatment, N_total, block_number = 4,
  drift = 0, simulation = 10000, conf_int = 0.95,
  alternative = "greater", correct = FALSE, replace = TRUE,
  early_stop = FALSE, method = "rosenberger", zvalue = c(1, 1.5, 2),
  rand_ratio = c(1, 1.5, 2, 2.5))

Arguments

p_control
scalar. Proportion of events under the control arm.

p_treatment
scalar. Proportion of events under the treatment arm.

N_total
scalar. Total sample size.

block_number
scalar. Number of blocks or time levels. The default is set to 4. If block_number is set to 1. This is a traditional RCT design.

drift
scalar. The increase or decrease in proportion of event over time. In this case, the proportion of failure changes in each block by the number of patient accured over the total sample size. The full drift effect is seen in the final block.

simulation
scalar. Number of simulation to be ran. The default is set to 10000.

conf_int
scalar. Confidence level of the interval.

alternative
character. A string specifying the alternative hypothesis, must be one of "less" or "greater" (default).

correct
logical. A logical indicating whether to apply continuity correction when computing the test statistic: one half is subtracted from all |O - E| differences; however, the correction will not be bigger than the differences themselves.

replace
logical. should sampling be with replacement? If replace is set to FALSE (default), the 0 for control, 1 for treatment is replicated to the closest integer and this vector is sampled with no replacement. If replace is set to TRUE, the sampling is done based on randomization ratio provided with replacement.

early_stop
logical. A logical indicating whether the trials are stopped early for success or futility.

method
character. The method used to alter randomization ratio. The function supports both rosenberg method and stepwise method.

zvalue
vector. The z-value cutoff for corrected chi-square test statistics. This parameter is ignored if the stepwise method is not used.

rand_ratio
vector. The randomization ratio is set based on the corrected chi-square test statistic. The length of rand_ratio should be the same length of zvalue. This parameter is ignored if the stepwise method is not used.

Value

a list with details on the simulation.

power

scalar. The power of the trial, ie. the proportion of success over the number of simulation ran.

p_control_estimate

scalar. The estimated proportion of events under the control group.

p_treatment_estimate

scalar. The estimated proportion of events under the treatment group.

N_enrolled

vector. The number of patients enrolled in the trial (sum of control and experimental group for each simulation. )

N_control

vector. The number of patients enrolled in the control group for each simulation.

N_control

vector. The number of patients enrolled in the experimental group for each simulation.

Examples

binomialfreq(p_control = 0.7, p_treatment = 0.65, N_total = 200,
            block_number = 2, simulation = 100)

#> $power
#> [1] 0
#> 
#> $p_control_estimate
#>   [1] 0.6893204 0.7156863 0.6562500 0.6764706 0.7333333 0.5957447 0.7477477
#>   [8] 0.8172043 0.7227723 0.7142857 0.7428571 0.7373737 0.7169811 0.7575758
#>  [15] 0.6761905 0.7419355 0.6888889 0.7079646 0.7019231 0.7608696 0.7171717
#>  [22] 0.7263158 0.6601942 0.7525773 0.6666667 0.7047619 0.6981132 0.7000000
#>  [29] 0.6481481 0.6947368 0.7872340 0.6388889 0.7926829 0.6116505 0.7129630
#>  [36] 0.7052632 0.6666667 0.7386364 0.6847826 0.7735849 0.6315789 0.7872340
#>  [43] 0.7346939 0.6862745 0.7227723 0.6702128 0.6800000 0.6862745 0.6914894
#>  [50] 0.7019231 0.6929825 0.7204301 0.7425743 0.6185567 0.7078652 0.7403846
#>  [57] 0.7128713 0.6634615 0.6938776 0.6800000 0.6698113 0.6699029 0.6880734
#>  [64] 0.7640449 0.6373626 0.7064220 0.6306306 0.7604167 0.7319588 0.7500000
#>  [71] 0.6666667 0.7272727 0.7263158 0.7586207 0.6607143 0.7415730 0.6960784
#>  [78] 0.6407767 0.6981132 0.7254902 0.7187500 0.7403846 0.7604167 0.7472527
#>  [85] 0.6804124 0.6451613 0.6530612 0.6095238 0.7600000 0.6831683 0.7623762
#>  [92] 0.6470588 0.6701031 0.7558140 0.7835052 0.6355140 0.7000000 0.6914894
#>  [99] 0.6590909 0.6600000
#> 
#> $p_treatment_estimate
#>   [1] 0.6804124 0.6020408 0.6250000 0.6122449 0.5818182 0.6603774 0.5842697
#>   [8] 0.6448598 0.6161616 0.6513761 0.6526316 0.6039604 0.6276596 0.6336634
#>  [15] 0.6842105 0.6168224 0.5909091 0.6551724 0.6770833 0.5925926 0.6237624
#>  [22] 0.6285714 0.6185567 0.7184466 0.6326531 0.7052632 0.6702128 0.6727273
#>  [29] 0.6304348 0.6476190 0.6320755 0.6521739 0.6101695 0.6804124 0.6739130
#>  [36] 0.6285714 0.6947368 0.5803571 0.5925926 0.6808511 0.5619048 0.6226415
#>  [43] 0.6274510 0.6734694 0.6363636 0.6886792 0.6800000 0.6530612 0.6509434
#>  [50] 0.6875000 0.7209302 0.6635514 0.6464646 0.6893204 0.6846847 0.6979167
#>  [57] 0.6060606 0.6666667 0.6176471 0.7100000 0.6063830 0.6701031 0.6593407
#>  [64] 0.6036036 0.6513761 0.6263736 0.6067416 0.6730769 0.6601942 0.6666667
#>  [71] 0.6265060 0.7623762 0.6000000 0.5929204 0.6136364 0.6126126 0.6428571
#>  [78] 0.6288660 0.6914894 0.6632653 0.6057692 0.6354167 0.6250000 0.6605505
#>  [85] 0.5631068 0.7196262 0.6470588 0.7157895 0.6600000 0.6565657 0.6060606
#>  [92] 0.6695652 0.6213592 0.7280702 0.6893204 0.6236559 0.6900000 0.6886792
#>  [99] 0.6696429 0.7100000
#> 
#> $prop_diff_estimate
#>   [1] -0.006421834 -0.113290633 -0.033589640 -0.066332665 -0.153379199
#>   [6]  0.053737374 -0.164377946 -0.173282289 -0.103870062 -0.065257166
#>  [11] -0.095384615 -0.134026805 -0.086650583 -0.120713856  0.007218769
#>  [16] -0.130680671 -0.112244898 -0.048785639 -0.025682183 -0.175850891
#>  [21] -0.088630439 -0.120154252 -0.043794147 -0.032711218 -0.036028823
#>  [26]  0.002010050 -0.029814666 -0.024489796 -0.017741935 -0.045956251
#>  [31] -0.153690187  0.011983471 -0.182401656  0.082946519 -0.036871961
#>  [36] -0.063674955  0.019487179 -0.157063594 -0.091220298 -0.103869654
#>  [41] -0.069179868 -0.165028665 -0.098268224 -0.012009608 -0.088980433
#>  [46]  0.016626115 -0.001610306 -0.036649215 -0.046028513 -0.016284680
#>  [51]  0.027947776 -0.057085427 -0.099616858  0.070070070 -0.025937183
#>  [56] -0.044088176 -0.106075797  0.009851226 -0.076230492  0.030000000
#>  [61] -0.062650602 -0.001614531 -0.028456105 -0.160153877  0.013510788
#>  [66] -0.080056463 -0.023081595 -0.089465219 -0.072822370 -0.082164329
#>  [71] -0.042671614  0.034206841 -0.125841322 -0.166157605 -0.052287582
#>  [76] -0.128366066 -0.054618474 -0.014732593 -0.001208703 -0.078644764
#>  [81] -0.114123837 -0.104208417 -0.146710257 -0.086781029 -0.116516517
#>  [86]  0.073969849 -0.004430125  0.105875276 -0.099639856 -0.021656306
#>  [91] -0.152195709  0.023589744 -0.049049049 -0.027743778 -0.091995947
#>  [96] -0.010502929 -0.010204082 -0.004866180  0.005091650  0.054347826
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1] 103 102  96 102  90  94 111  93 101  91 105  99 106  99 105  93  90 113
#>  [19] 104  92  99  95 103  97 102 105 106  90 108  95  94 108  82 103 108  95
#>  [37] 105  88  92 106  95  94  98 102 101  94 100 102  94 104 114  93 101  97
#>  [55]  89 104 101 104  98 100 106 103 109  89  91 109 111  96  97 104 117  99
#>  [73]  95  87 112  89 102 103 106 102  96 104  96  91  97  93  98 105 100 101
#>  [91] 101  85  97  86  97 107 100  94  88 100
#> 
#> $N_treatment
#>   [1]  97  98 104  98 110 106  89 107  99 109  95 101  94 101  95 107 110  87
#>  [19]  96 108 101 105  97 103  98  95  94 110  92 105 106  92 118  97  92 105
#>  [37]  95 112 108  94 105 106 102  98  99 106 100  98 106  96  86 107  99 103
#>  [55] 111  96  99  96 102 100  94  97  91 111 109  91  89 104 103  96  83 101
#>  [73] 105 113  88 111  98  97  94  98 104  96 104 109 103 107 102  95 100  99
#>  [91]  99 115 103 114 103  93 100 106 112 100
#> 

binomialfreq(p_control = 0.5, p_treatment = 0.40, N_total = 200,
            block_number = 2, simulation = 100, drift = -0.15)

#> $power
#> [1] 0
#> 
#> $p_control_estimate
#>   [1] 0.4285714 0.5393258 0.4204545 0.3333333 0.4019608 0.4761905 0.4574468
#>   [8] 0.3578947 0.3644860 0.3191489 0.4181818 0.2959184 0.3921569 0.3132530
#>  [15] 0.3666667 0.2765957 0.4042553 0.3541667 0.4193548 0.3936170 0.3867925
#>  [22] 0.3883495 0.4315789 0.4479167 0.3269231 0.3369565 0.4313725 0.3883495
#>  [29] 0.3440860 0.3369565 0.4333333 0.4077670 0.4270833 0.4021739 0.3222222
#>  [36] 0.4242424 0.3846154 0.3736264 0.4000000 0.3628319 0.3434343 0.3917526
#>  [43] 0.4111111 0.3333333 0.3636364 0.3750000 0.3333333 0.3461538 0.3750000
#>  [50] 0.5679012 0.3333333 0.3522727 0.4100000 0.3478261 0.4897959 0.4022989
#>  [57] 0.4205607 0.4222222 0.4300000 0.4086022 0.3958333 0.3888889 0.4095238
#>  [64] 0.3838384 0.4361702 0.4864865 0.3333333 0.3838384 0.4259259 0.3291139
#>  [71] 0.3586957 0.3944954 0.3764706 0.4019608 0.3434343 0.4653465 0.4200000
#>  [78] 0.3404255 0.3297872 0.4150943 0.4767442 0.3900000 0.4888889 0.4536082
#>  [85] 0.4216867 0.3177570 0.4095238 0.4100000 0.4158416 0.3627451 0.3263158
#>  [92] 0.3921569 0.3020833 0.3243243 0.3457944 0.3908046 0.3235294 0.4252874
#>  [99] 0.4090909 0.4395604
#> 
#> $p_treatment_estimate
#>   [1] 0.2568807 0.2162162 0.2410714 0.3545455 0.2346939 0.2241379 0.3113208
#>   [8] 0.3523810 0.3655914 0.2452830 0.2666667 0.3039216 0.3265306 0.2905983
#>  [15] 0.2909091 0.3490566 0.2924528 0.2692308 0.3177570 0.3396226 0.2234043
#>  [22] 0.2577320 0.2380952 0.2307692 0.3020833 0.2962963 0.2959184 0.2989691
#>  [29] 0.2523364 0.3055556 0.3000000 0.2989691 0.1923077 0.3055556 0.3545455
#>  [36] 0.3069307 0.2213115 0.3486239 0.2400000 0.3103448 0.2178218 0.2815534
#>  [43] 0.3363636 0.2920354 0.3000000 0.2500000 0.3367347 0.3125000 0.3035714
#>  [50] 0.2857143 0.2884615 0.3303571 0.2400000 0.2314815 0.2647059 0.2743363
#>  [57] 0.2688172 0.2818182 0.3300000 0.2429907 0.2692308 0.2826087 0.3263158
#>  [64] 0.3267327 0.3773585 0.3932584 0.3052632 0.2574257 0.2608696 0.2561983
#>  [71] 0.2407407 0.2747253 0.2173913 0.2959184 0.3267327 0.3232323 0.2700000
#>  [78] 0.3207547 0.2735849 0.3085106 0.2807018 0.3500000 0.2454545 0.3398058
#>  [85] 0.2051282 0.3225806 0.2631579 0.3600000 0.3636364 0.3061224 0.2571429
#>  [92] 0.2448980 0.3076923 0.3370787 0.3333333 0.2743363 0.2448980 0.3628319
#>  [99] 0.3035714 0.2201835
#> 
#> $prop_diff_estimate
#>   [1] -0.161749536 -0.324861652 -0.179383117  0.015656771 -0.167266907
#>   [6] -0.245065789 -0.137592138 -0.031794872  0.005111429 -0.066881547
#>  [11] -0.145285310  0.014056225 -0.060921844 -0.025810448 -0.074666127
#>  [16]  0.067350221 -0.108835341 -0.095218635 -0.092777213 -0.057211926
#>  [21] -0.165461847 -0.144478217 -0.192454345 -0.211602671 -0.016666667
#>  [26] -0.035353535 -0.126452906 -0.092766981 -0.086119554 -0.027419355
#>  [31] -0.130962005 -0.107393308 -0.217445743 -0.112745098  0.030315279
#>  [36] -0.117823565 -0.160759494 -0.023814329 -0.154630763 -0.053145999
#>  [41] -0.123417085 -0.111729540 -0.069387755 -0.042115972 -0.065805658
#>  [46] -0.123647295  0.007450665 -0.023876973 -0.069444444 -0.274869656
#>  [51] -0.044871795 -0.021112464 -0.166399037 -0.104943274 -0.212206952
#>  [56] -0.130523315 -0.129744799 -0.144897959 -0.100000000 -0.169258736
#>  [61] -0.126016260 -0.103030303 -0.083885210 -0.043960476 -0.043788187
#>  [66] -0.090927542 -0.030150754 -0.120185521 -0.154747475 -0.074554974
#>  [71] -0.093841032 -0.119003878 -0.133458255 -0.107630522 -0.014237016
#>  [76] -0.142367413 -0.149939783 -0.014877362 -0.054582485 -0.106464646
#>  [81] -0.187784679 -0.043269231 -0.230994152 -0.114114114 -0.215653965
#>  [86]  0.009052505 -0.150110375 -0.048374147 -0.045372051 -0.061248999
#>  [91] -0.066693532 -0.128536783 -0.006451613  0.018844985 -0.004432803
#>  [96] -0.115157681 -0.079518072 -0.074334398 -0.099796334 -0.219047619
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1]  91  89  88  90 102  84  94  95 107  94 110  98 102  83  90  94  94  96
#>  [19]  93  94 106 103  95  96 104  92 102 103  93  92  90 103  96  92  90  99
#>  [37]  78  91 100 113  99  97  90  87 110 104 102 104  88  81  96  88 100  92
#>  [55]  98  87 107  90 100  93  96 108 105  99  94 111 105  99 108  79  92 109
#>  [73]  85 102  99 101 100  94  94 106  86 100  90  97  83 107 105 100 101 102
#>  [91]  95 102  96 111 107  87 102  87  88  91
#> 
#> $N_treatment
#>   [1] 109 111 112 110  98 116 106 105  93 106  90 102  98 117 110 106 106 104
#>  [19] 107 106  94  97 105 104  96 108  98  97 107 108 110  97 104 108 110 101
#>  [37] 122 109 100  87 101 103 110 113  90  96  98  96 112 119 104 112 100 108
#>  [55] 102 113  93 110 100 107 104  92  95 101 106  89  95 101  92 121 108  91
#>  [73] 115  98 101  99 100 106 106  94 114 100 110 103 117  93  95 100  99  98
#>  [91] 105  98 104  89  93 113  98 113 112 109
#>