Block Design for Response-Adaptive Randomization

Simulation for normal mean for block design for response-adaptive randomization with time as a confounding

normalRAR(mu_control, mu_treatment, sd_control, sd_treatment, N_total,
  block_number = 4, simulation = 10000, zvalue = NULL,
  rand_ratio = NULL, conf_int = 0.95, alternative = "less",
  replace = FALSE)

Arguments

mu_control
scalar. Mean outcome in the control arm.

mu_treatment
scalar. Mean outcome in the treatment arm.

sd_control
scalar. Standard deviation of outcome in the control arm.

sd_treatment
scalar. Standard deviation of outcome in the treatment

N_total
scalar. Total sample size.

block_number
scalar. Number of blocks or time levels. The default is set to 4. If block_number is set to 1. This is a traditional RCT design.

simulation
scalar. Number of simulation to be ran. The default is set to 10000.

zvalue
vector. The z-value cutoff for corrected chi-square test statistics.

rand_ratio
vector. The randomization ratio is set based on the corrected chi-square test statistic. The length of rand_ratio should be the same length of zvalue.

conf_int
scalar. Confidence level of the interval.

alternative
character. A string specifying the alternative hypothesis, must be one of "less" (default) or "greater".

replace
character. should sampling be with replacement? If replace is set to FALSE (default), the 0 for control, 1 for treatment is replicated to the closest integer and this vector is sampled with no replacement. If replace is set to TRUE, the sampling is done based on randomization ratio provided with replacement.

Value

a list with details on the simulation.

power

scalar. The power of the trial, ie. the proportion of success over the number of simulation ran.

N_enrolled

vector. The number of patients enrolled in the trial (sum of control and experimental group for each simulation. )

N_control

vector. The number of patients enrolled in the control group for each simulation.

N_control

vector. The number of patients enrolled in the experimental group for each simulation.

Examples

normalRAR(mu_control = 5.7, mu_treatment = 0.65,
          sd_control = 1.1, sd_treatment = 1.2,
          N_total = 200, block_number = 2, simulation = 100)

#> $power
#> [1] 0
#> 
#> $N_enrolled
#> NULL
#> 
#> $N_control
#> NULL
#> 
#> $N_treatment
#> NULL
#>