Block Design for Response-Adaptive Randomization for Binomial Data

Simulation for binomial counts for block design for response-adaptive randomization with time as a confounding

binomialbayes(p_control, p_treatment, N_total, block_number = 4,
  drift = 0, simulation = 10000, a0 = 0.5, b0 = 0.5, p = 0.5,
  number_mcmc = 10000, prob_accept_ha = 0.95,
  early_success_prob = 0.99, futility_prob = 0.01,
  alternative = "greater")

Arguments

p_control
scalar. Proportion of events under the control arm.

p_treatment
scalar. Proportion of events under the treatment arm.

N_total
scalar. Total sample size.

block_number
scalar. Number of blocks or time levels. The default is set to 4. If block_number is set to 1. This is a traditional RCT design.

drift
scalar. The increase or decrease in proportion of event over time. In this case, the proportion of failure changes in each block by the number of patient accured over the total sample size. The full drift effect is seen in the final block.

simulation
scalar. Number of simulation to be ran. The default is set to 10000.

a0
scalar. Prior value for the beta rate Beta(a0, b0). Default is 0.5.

b0
scalar. Prior value for the beta rate Beta(a0, b0). Default is 0.5.

p
sclar. Power for randomization ratio.

number_mcmc
scalar. Number of Monte Carlo Markov Chain draws in sampling posterior.

prob_accept_ha
scalar. Probability of accepting alternative hypothesis.

early_success_prob
scalar. Probability of stopping early for success.

futility_prob
scalar. Probability of stopping early for futility.

alternative
character. A string specifying the alternative hypothesis, must be one of "less" (default) or "greater".

Value

a list with details on the simulation.

power

scalar. The power of the trial, ie. the proportion of success over the number of simulation ran.

p_control_estimate

scalar. The estimated proportion of events under the control group.

p_treatment_estimate

scalar. The estimated proportion of events under the treatment group.

N_enrolled

vector. The number of patients enrolled in the trial (sum of control and experimental group for each simulation. )

N_control

vector. The number of patients enrolled in the control group for each simulation.

N_control

vector. The number of patients enrolled in the experimental group for each simulation.

Examples

binomialbayes(p_control = 0.20, p_treatment = 0.30, N_total = 100, simulation = 10)

#> $power
#> [1] 0.3
#> 
#> $prop_diff_estimate
#>  [1]  0.2922111 -0.2827446  0.3917874  1.2484785  0.2937040  1.5931985
#>  [7] -0.4140110  0.7439645  1.0589844 -0.2072926
#> 
#> $N_enrolled
#>  [1] 100 100 100  50 100 100 100 100 100 100
#> 
#> $N_control
#>  [1] 37 52 48 25 36 33 65 34 46 66
#> 
#> $N_treatment
#>  [1] 63 48 52 25 64 67 35 66 54 34
#> 
#> $early_success
#>  [1] 0 0 0 1 0 0 0 0 0 0
#> 
#> $early_futilty
#>  [1] 0 0 0 0 0 0 0 0 0 0
#>