Block Design for Response-Adaptive Randomization for Binomial Data

Simulation for binomial counts for block design for response-adaptive randomization with time as a confounding

binomialfreq(p_control, p_treatment, N_total, block_number = 4,
  drift = 0, simulation = 10000, conf_int = 0.95,
  alternative = "greater", correct = FALSE, replace = TRUE,
  early_stop = FALSE, method = "rosenberger", zvalue = c(1, 1.5, 2),
  rand_ratio = c(1, 1.5, 2, 2.5))

Arguments

p_control
scalar. Proportion of events under the control arm.

p_treatment
scalar. Proportion of events under the treatment arm.

N_total
scalar. Total sample size.

block_number
scalar. Number of blocks or time levels. The default is set to 4. If block_number is set to 1. This is a traditional RCT design.

drift
scalar. The increase or decrease in proportion of event over time. In this case, the proportion of failure changes in each block by the number of patient accured over the total sample size. The full drift effect is seen in the final block.

simulation
scalar. Number of simulation to be ran. The default is set to 10000.

conf_int
scalar. Confidence level of the interval.

alternative
character. A string specifying the alternative hypothesis, must be one of "less" or "greater" (default).

correct
logical. A logical indicating whether to apply continuity correction when computing the test statistic: one half is subtracted from all |O - E| differences; however, the correction will not be bigger than the differences themselves.

replace
logical. should sampling be with replacement? If replace is set to FALSE (default), the 0 for control, 1 for treatment is replicated to the closest integer and this vector is sampled with no replacement. If replace is set to TRUE, the sampling is done based on randomization ratio provided with replacement.

early_stop
logical. A logical indicating whether the trials are stopped early for success or futility.

method
character. The method used to alter randomization ratio. The function supports both rosenberg method and stepwise method.

zvalue
vector. The z-value cutoff for corrected chi-square test statistics. This parameter is ignored if the stepwise method is not used.

rand_ratio
vector. The randomization ratio is set based on the corrected chi-square test statistic. The length of rand_ratio should be the same length of zvalue. This parameter is ignored if the stepwise method is not used.

Value

a list with details on the simulation.

power

scalar. The power of the trial, ie. the proportion of success over the number of simulation ran.

p_control_estimate

scalar. The estimated proportion of events under the control group.

p_treatment_estimate

scalar. The estimated proportion of events under the treatment group.

N_enrolled

vector. The number of patients enrolled in the trial (sum of control and experimental group for each simulation. )

N_control

vector. The number of patients enrolled in the control group for each simulation.

N_control

vector. The number of patients enrolled in the experimental group for each simulation.

Examples

binomialfreq(p_control = 0.7, p_treatment = 0.65, N_total = 200,
            block_number = 2, simulation = 100)

#> $power
#> [1] 0.04
#> 
#> $p_control_estimate
#>   [1] 0.7184466 0.6597938 0.6542056 0.6964286 0.7582418 0.7184466 0.7171717
#>   [8] 0.7555556 0.7741935 0.6300000 0.6853933 0.6559140 0.6990291 0.7395833
#>  [15] 0.7473684 0.7244898 0.7300000 0.6698113 0.6666667 0.6923077 0.6831683
#>  [22] 0.6595745 0.6923077 0.6601942 0.6489362 0.6842105 0.6800000 0.7244898
#>  [29] 0.6703297 0.6380952 0.7956989 0.6944444 0.7127660 0.6571429 0.7386364
#>  [36] 0.6454545 0.6759259 0.7700000 0.6800000 0.7578947 0.6600000 0.6770833
#>  [43] 0.7216495 0.6836735 0.7378641 0.6700000 0.6074766 0.6818182 0.7446809
#>  [50] 0.7528090 0.6770833 0.7241379 0.6947368 0.7272727 0.7281553 0.6568627
#>  [57] 0.6179775 0.6808511 0.6666667 0.6888889 0.5384615 0.6941176 0.7692308
#>  [64] 0.6813187 0.7079646 0.7526882 0.6991150 0.7289720 0.7500000 0.6407767
#>  [71] 0.6276596 0.6500000 0.7281553 0.7821782 0.7500000 0.6761905 0.7476636
#>  [78] 0.7111111 0.6896552 0.7378641 0.6091954 0.6666667 0.6315789 0.7717391
#>  [85] 0.5698925 0.7087379 0.7065217 0.6930693 0.7128713 0.5445545 0.6938776
#>  [92] 0.7156863 0.6923077 0.7029703 0.7722772 0.6966292 0.6952381 0.7526882
#>  [99] 0.7647059 0.7272727
#> 
#> $p_treatment_estimate
#>   [1] 0.6804124 0.7087379 0.6774194 0.6818182 0.7155963 0.5567010 0.6831683
#>   [8] 0.6454545 0.7102804 0.7400000 0.5225225 0.6822430 0.6185567 0.6346154
#>  [15] 0.6571429 0.5784314 0.6200000 0.7340426 0.6915888 0.6458333 0.6767677
#>  [22] 0.6792453 0.5963303 0.7319588 0.6415094 0.6476190 0.6200000 0.6274510
#>  [29] 0.6513761 0.6315789 0.6542056 0.6413043 0.7264151 0.6842105 0.6607143
#>  [36] 0.6555556 0.6739130 0.6500000 0.6000000 0.6952381 0.6000000 0.6634615
#>  [43] 0.6796117 0.7156863 0.6288660 0.6300000 0.6559140 0.5888889 0.7075472
#>  [50] 0.6666667 0.6153846 0.6194690 0.7333333 0.6633663 0.6907216 0.6428571
#>  [57] 0.6846847 0.6509434 0.6966292 0.6818182 0.7247706 0.6695652 0.6972477
#>  [64] 0.6422018 0.7586207 0.7289720 0.7241379 0.7311828 0.6442308 0.6082474
#>  [71] 0.5754717 0.6666667 0.6494845 0.7171717 0.6388889 0.6421053 0.6666667
#>  [78] 0.6000000 0.6106195 0.7216495 0.7256637 0.7596154 0.6571429 0.6388889
#>  [85] 0.5887850 0.6597938 0.6759259 0.6464646 0.7070707 0.6868687 0.6470588
#>  [92] 0.6224490 0.6238532 0.7272727 0.5959596 0.6306306 0.6421053 0.6355140
#>  [99] 0.7244898 0.6138614
#> 
#> $prop_diff_estimate
#>   [1] -0.046630014  0.048048048  0.023115578 -0.014610390 -0.044803229
#>   [6] -0.166767250 -0.031631632 -0.103811841 -0.063919598  0.109043617
#>  [11] -0.162583519  0.024542346 -0.079479479 -0.101612903 -0.093491840
#>  [16] -0.142971888 -0.109843938  0.065672844  0.023977433 -0.039837398
#>  [21] -0.006001200  0.021310816 -0.094198504  0.072446317 -0.011201629
#>  [26] -0.019607843 -0.060424170 -0.094987874 -0.017154390 -0.008205128
#>  [31] -0.139007092 -0.053140097  0.013253012  0.033846154 -0.077344702
#>  [36]  0.010137875 -0.002012882 -0.109577922 -0.080000000 -0.062311558
#>  [41] -0.059294872 -0.013643660 -0.041139876  0.032865731 -0.106992587
#>  [46] -0.038940185  0.051498692 -0.092270336 -0.031829170 -0.084838644
#>  [51] -0.055287495 -0.105908812  0.038299579 -0.062212442 -0.035663627
#>  [56] -0.017317761  0.068398620  0.012033195  0.033083012 -0.013787510
#>  [61]  0.202082058 -0.024615385 -0.071788667 -0.039215686  0.039958805
#>  [66] -0.021927178  0.023047114  0.001809045 -0.103136309 -0.037126229
#>  [71] -0.052525253  0.015423093 -0.083555191 -0.065665666 -0.111111111
#>  [76] -0.033687588 -0.089039687 -0.110300081 -0.078802688 -0.015015015
#>  [81]  0.116786818  0.091983968  0.030769231 -0.144367629  0.015917792
#>  [86] -0.052895211 -0.029173420 -0.044244244 -0.008643216  0.141167034
#>  [91] -0.046818727 -0.093237295 -0.067352289  0.023404681 -0.174974975
#>  [96] -0.056436663 -0.052578768 -0.100592926 -0.047696039 -0.112422484
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1] 103  97 107 112  91 103  99  90  93 100  89  93 103  96  95  98 100 106
#>  [19]  93 104 101  94  91 103  94  95 100  98  91 105  93 108  94 105  88 110
#>  [37] 108 100 100  95 100  96  97  98 103 100 107 110  94  89  96  87  95  99
#>  [55] 103 102  89  94 111  90  91  85  91  91 113  93 113 107  96 103  94  80
#>  [73] 103 101  92 105 107  90  87 103  87  96  95  92  93 103  92 101 101 101
#>  [91]  98 102  91 101 101  89 105  93 102  99
#> 
#> $N_treatment
#>   [1]  97 103  93  88 109  97 101 110 107 100 111 107  97 104 105 102 100  94
#>  [19] 107  96  99 106 109  97 106 105 100 102 109  95 107  92 106  95 112  90
#>  [37]  92 100 100 105 100 104 103 102  97 100  93  90 106 111 104 113 105 101
#>  [55]  97  98 111 106  89 110 109 115 109 109  87 107  87  93 104  97 106 120
#>  [73]  97  99 108  95  93 110 113  97 113 104 105 108 107  97 108  99  99  99
#>  [91] 102  98 109  99  99 111  95 107  98 101
#> 

binomialfreq(p_control = 0.5, p_treatment = 0.40, N_total = 200,
            block_number = 2, simulation = 100, drift = -0.15)

#> $power
#> [1] 0
#> 
#> $p_control_estimate
#>   [1] 0.4128440 0.4210526 0.3431373 0.2980769 0.3900000 0.3877551 0.3707865
#>   [8] 0.3673469 0.3956044 0.3863636 0.4022989 0.4021739 0.3737374 0.4795918
#>  [15] 0.4186047 0.4339623 0.3366337 0.3723404 0.4193548 0.4044944 0.3796296
#>  [22] 0.3863636 0.4526316 0.4042553 0.3368421 0.3619048 0.3953488 0.3936170
#>  [29] 0.4893617 0.4747475 0.4130435 0.3711340 0.4255319 0.3800000 0.4368932
#>  [36] 0.3904762 0.3723404 0.4526316 0.3516484 0.3900000 0.4065934 0.5106383
#>  [43] 0.3888889 0.4324324 0.4347826 0.4318182 0.4845361 0.3854167 0.4230769
#>  [50] 0.5300000 0.3861386 0.3870968 0.3191489 0.3333333 0.3809524 0.4252874
#>  [57] 0.3596491 0.3932584 0.3723404 0.3679245 0.3936170 0.3482143 0.3295455
#>  [64] 0.3939394 0.3809524 0.3854167 0.3535354 0.3333333 0.4526316 0.4186047
#>  [71] 0.3800000 0.3711340 0.3510638 0.3888889 0.3047619 0.3164557 0.4375000
#>  [78] 0.3302752 0.3333333 0.4183673 0.3750000 0.2604167 0.3917526 0.4111111
#>  [85] 0.4019608 0.3797468 0.3707865 0.4361702 0.5222222 0.3644860 0.3777778
#>  [92] 0.3555556 0.4193548 0.3518519 0.3636364 0.4285714 0.3917526 0.3232323
#>  [99] 0.3921569 0.4705882
#> 
#> $p_treatment_estimate
#>   [1] 0.3516484 0.2952381 0.3469388 0.3125000 0.3800000 0.3235294 0.3153153
#>   [8] 0.2843137 0.3119266 0.2500000 0.2566372 0.2500000 0.3069307 0.2450980
#>  [15] 0.2982456 0.3617021 0.3535354 0.2735849 0.2242991 0.2702703 0.2717391
#>  [22] 0.2767857 0.3047619 0.1886792 0.2476190 0.3473684 0.2105263 0.2830189
#>  [29] 0.2641509 0.2475248 0.3148148 0.3398058 0.2358491 0.2800000 0.2886598
#>  [36] 0.2842105 0.2641509 0.2761905 0.2660550 0.3600000 0.3119266 0.2641509
#>  [43] 0.3181818 0.2539683 0.2222222 0.2589286 0.2912621 0.3365385 0.2540984
#>  [50] 0.2700000 0.2828283 0.2429907 0.2547170 0.2363636 0.2736842 0.3008850
#>  [57] 0.3023256 0.2432432 0.3490566 0.3191489 0.3207547 0.2840909 0.2857143
#>  [64] 0.2574257 0.3368421 0.2692308 0.2574257 0.3461538 0.2380952 0.2719298
#>  [71] 0.3000000 0.3495146 0.2547170 0.2934783 0.2421053 0.3305785 0.2596154
#>  [78] 0.2637363 0.3461538 0.2941176 0.2916667 0.3365385 0.2718447 0.3363636
#>  [85] 0.2653061 0.2644628 0.2612613 0.3301887 0.2272727 0.2688172 0.2818182
#>  [92] 0.3454545 0.3177570 0.3260870 0.3069307 0.3137255 0.2621359 0.2376238
#>  [99] 0.2959184 0.3775510
#> 
#> $prop_diff_estimate
#>   [1] -0.038375179 -0.127433273  0.003801521  0.004032258 -0.012820513
#>   [6] -0.063126253 -0.062715648 -0.084867894 -0.083686227 -0.130971993
#>  [11] -0.135163675 -0.129194631 -0.069269269 -0.227755102 -0.115651819
#>  [16] -0.073493976  0.013613614 -0.099156118 -0.194773869 -0.139638725
#>  [21] -0.107890499 -0.111788618 -0.135835203 -0.203568532 -0.089833567
#>  [26] -0.011082007 -0.180544106 -0.115199035 -0.208888889 -0.225527638
#>  [31] -0.101260675 -0.026139818 -0.182514102 -0.100040016 -0.149149149
#>  [36] -0.102512563 -0.108189482 -0.156281713 -0.081723928 -0.030612245
#>  [41] -0.094427558 -0.228710462 -0.091430914 -0.178230703 -0.193627451
#>  [46] -0.172049689 -0.193193193 -0.044141252 -0.166523605 -0.253623188
#>  [51] -0.103620724 -0.148498892 -0.065863454 -0.091240876 -0.109208140
#>  [56] -0.124512821 -0.060408163 -0.129329781 -0.015151515 -0.047389558
#>  [61] -0.072862304 -0.064123377 -0.041666667 -0.131531532 -0.039333735
#>  [66] -0.097861842 -0.100862242  0.012903226 -0.209015565 -0.133638634
#>  [71] -0.082032813 -0.021021021 -0.095315682 -0.095161290 -0.063765791
#>  [76]  0.015078534 -0.166531769 -0.062361251  0.015248796 -0.119238477
#>  [81] -0.083333333  0.075350701 -0.108173562 -0.071980480 -0.133306645
#>  [86] -0.109293056 -0.102938197 -0.118665019 -0.280708986 -0.103748734
#>  [91] -0.092051906 -0.016592473 -0.103802052 -0.026178010 -0.013505218
#>  [96] -0.109943411 -0.124318870 -0.088156540 -0.075990299 -0.076126675
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1] 109  95 102 104 100  98  89  98  91  88  87  92  99  98  86 106 101  94
#>  [19]  93  89 108  88  95  94  95 105  86  94  94  99  92  97  94 100 103 105
#>  [37]  94  95  91 100  91  94  90  74  92  88  97  96  78 100 101  93  94  90
#>  [55] 105  87 114  89  94 106  94 112  88  99 105  96  99  96  95  86 100  97
#>  [73]  94 108 105  79  96 109  96  98 104  96  97  90 102  79  89  94  90 107
#>  [91]  90  90  93 108  99  98  97  99 102 102
#> 
#> $N_treatment
#>   [1]  91 105  98  96 100 102 111 102 109 112 113 108 101 102 114  94  99 106
#>  [19] 107 111  92 112 105 106 105  95 114 106 106 101 108 103 106 100  97  95
#>  [37] 106 105 109 100 109 106 110 126 108 112 103 104 122 100  99 107 106 110
#>  [55]  95 113  86 111 106  94 106  88 112 101  95 104 101 104 105 114 100 103
#>  [73] 106  92  95 121 104  91 104 102  96 104 103 110  98 121 111 106 110  93
#>  [91] 110 110 107  92 101 102 103 101  98  98
#>