Block Design for Response-Adaptive Randomization for Binomial Data

Simulation for binomial counts for block design for response-adaptive randomization with time as a confounding

binomialfreq(p_control, p_treatment, N_total, block_number = 4,
  drift = 0, simulation = 10000, conf_int = 0.95,
  alternative = "greater", correct = FALSE, replace = TRUE,
  early_stop = FALSE, size_equal_randomization = 20,
  min_patient_earlystop = 20)

Arguments

p_control
scalar. Proportion of events under the control arm.

p_treatment
scalar. Proportion of events under the treatment arm.

N_total
scalar. Total sample size.

block_number
scalar. Number of blocks or time levels. The default is set to 4. If block_number is set to 1. This is a traditional RCT design.

drift
scalar. The increase or decrease in proportion of event over time. In this case, the proportion of failure changes in each block by the number of patient accured over the total sample size. The full drift effect is seen in the final block.

simulation
scalar. Number of simulation to be ran. The default is set to 10000.

conf_int
scalar. Confidence level of the interval.

alternative
character. A string specifying the alternative hypothesis, must be one of "less" or "greater" (default).

correct
logical. A logical indicating whether to apply continuity correction when computing the test statistic: one half is subtracted from all |O - E| differences; however, the correction will not be bigger than the differences themselves.

replace
logical. should sampling be with replacement? If replace is set to FALSE (default), the 0 for control, 1 for treatment is replicated to the closest integer and this vector is sampled with no replacement. If replace is set to TRUE, the sampling is done based on randomization ratio provided with replacement.

early_stop
logical. A logical indicating whether the trials are stopped early for success or futility.

size_equal_randomization
scalar. The number of run in patients because adaptive randomization is applied.

min_patient_earlystop
scalar. Minimum number of patients before early stopping rule is applied.

Value

a list with details on the simulation.

power

scalar. The power of the trial, ie. the proportion of success over the number of simulation ran.

p_control_estimate

scalar. The estimated proportion of events under the control group.

p_treatment_estimate

scalar. The estimated proportion of events under the treatment group.

N_enrolled

vector. The number of patients enrolled in the trial (sum of control and experimental group for each simulation. )

N_control

vector. The number of patients enrolled in the control group for each simulation.

N_control

vector. The number of patients enrolled in the experimental group for each simulation.

randomization_ratio

matrix. The randomization ratio allocated for each block.

Examples

binomialfreq(p_control = 0.7, p_treatment = 0.65, N_total = 200,
            block_number = 2, simulation = 100)

#> $power
#> [1] 0.01
#> 
#> $p_control_estimate
#>   [1] 0.7450980 0.7100000 0.7115385 0.7700000 0.7500000 0.6831683 0.7281553
#>   [8] 0.6864407 0.7142857 0.7300000 0.6448598 0.8173077 0.6272727 0.7586207
#>  [15] 0.7340426 0.7131148 0.7227723 0.7075472 0.6990291 0.7058824 0.7010309
#>  [22] 0.6481481 0.6595745 0.6666667 0.7647059 0.6853933 0.7314815 0.7142857
#>  [29] 0.6982759 0.6813187 0.7363636 0.7663551 0.7395833 0.6969697 0.6629213
#>  [36] 0.6703297 0.7173913 0.6881720 0.6804124 0.7157895 0.7383178 0.6732673
#>  [43] 0.6836735 0.6603774 0.6836735 0.7096774 0.7033898 0.6800000 0.6893204
#>  [50] 0.7238095 0.7254902 0.7326733 0.6321839 0.7010309 0.7254902 0.6161616
#>  [57] 0.5900000 0.6041667 0.6810345 0.7291667 0.6796117 0.7254902 0.7500000
#>  [64] 0.7017544 0.7222222 0.7288136 0.6808511 0.6886792 0.6730769 0.6886792
#>  [71] 0.6666667 0.5494505 0.6601942 0.6836735 0.7948718 0.7171717 0.7428571
#>  [78] 0.7800000 0.6421053 0.6804124 0.7435897 0.6767677 0.6875000 0.7272727
#>  [85] 0.6893204 0.7525773 0.6893204 0.6161616 0.7244898 0.7311828 0.6703297
#>  [92] 0.7500000 0.7087379 0.7428571 0.6930693 0.7542373 0.7356322 0.6272727
#>  [99] 0.7387387 0.6638655
#> 
#> $p_treatment_estimate
#>   [1] 0.7142857 0.5900000 0.6458333 0.6200000 0.6100000 0.6060606 0.5567010
#>   [8] 0.6951220 0.5135135 0.6600000 0.7204301 0.7291667 0.5888889 0.6309524
#>  [15] 0.6132075 0.6666667 0.6161616 0.6914894 0.6597938 0.6632653 0.6213592
#>  [22] 0.6413043 0.6226415 0.5760870 0.6530612 0.6936937 0.6847826 0.6477273
#>  [29] 0.5714286 0.6238532 0.7555556 0.5806452 0.6923077 0.7128713 0.6486486
#>  [36] 0.6513761 0.7222222 0.7009346 0.6019417 0.5619048 0.6451613 0.5858586
#>  [43] 0.6078431 0.6808511 0.7058824 0.6635514 0.6463415 0.6100000 0.7319588
#>  [50] 0.6631579 0.6530612 0.5959596 0.5929204 0.6310680 0.6530612 0.6831683
#>  [57] 0.6600000 0.5961538 0.6666667 0.6826923 0.7010309 0.5612245 0.6562500
#>  [64] 0.6162791 0.6847826 0.6829268 0.6132075 0.5425532 0.6145833 0.6595745
#>  [71] 0.7211538 0.7798165 0.6494845 0.6666667 0.5903614 0.6336634 0.6736842
#>  [78] 0.6000000 0.6476190 0.6116505 0.6024096 0.6138614 0.6730769 0.6534653
#>  [85] 0.6494845 0.7669903 0.6804124 0.6534653 0.6274510 0.7383178 0.6972477
#>  [92] 0.6700000 0.6804124 0.6736842 0.6666667 0.6097561 0.7079646 0.6111111
#>  [99] 0.6067416 0.5925926
#> 
#> $prop_diff_estimate
#>   [1] -0.034538153 -0.121794872 -0.069561721 -0.151060424 -0.139423077
#>   [6] -0.077015403 -0.181022195  0.008681273 -0.200806813 -0.064259486
#>  [11]  0.081599677 -0.081825493 -0.040016168 -0.125051504 -0.119277108
#>  [16] -0.050315789 -0.107916582 -0.020104544 -0.052270434 -0.044457978
#>  [21] -0.078395439 -0.011996747 -0.035858179 -0.101212121 -0.110843373
#>  [26]  0.007700101 -0.042139384 -0.066788469 -0.130407911 -0.057517659
#>  [31]  0.018042306 -0.163113227 -0.048296593  0.014414414 -0.014982790
#>  [36] -0.024615385  0.007258065  0.005086470 -0.087762167 -0.161512027
#>  [41] -0.087255100 -0.089246231 -0.079628133  0.018094089  0.020240481
#>  [46] -0.042848521 -0.054771784 -0.061403509  0.043277900 -0.064745281
#>  [51] -0.074148297 -0.129648241 -0.043282051 -0.063895146 -0.072457966
#>  [56]  0.066013203  0.068108974 -0.008416834 -0.013757700 -0.046292585
#>  [61]  0.025227043 -0.154853000 -0.099918267 -0.080287474 -0.040606061
#>  [66] -0.047619048 -0.067068273 -0.160602258 -0.057383628 -0.028112450
#>  [71]  0.047956293  0.233177475 -0.004425669  0.001042101 -0.202492212
#>  [76] -0.074687948 -0.071558520 -0.173737374  0.007035176 -0.059334006
#>  [81] -0.142238714 -0.062612523 -0.014423077 -0.073873874 -0.040640641
#>  [86]  0.016229213 -0.009009009  0.029441420 -0.108135320  0.011686480
#>  [91]  0.030386157 -0.078787879 -0.025045081 -0.068979346 -0.025605121
#>  [96] -0.147034252 -0.029777687 -0.078982010 -0.125534079 -0.071384105
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1] 102 100 104 100 100 101 103 118 126 100 107 104 110 116  94 122 101 106
#>  [19] 103 102  97 108  94 108 102  89 108 112 116  91 110 107  96  99  89  91
#>  [37]  92  93  97  95 107 101  98 106  98  93 118 100 103 105 102 101  87  97
#>  [55] 102  99 100  96 116  96 103 102 104 114 108 118  94 106 104 106  96  91
#>  [73] 103  98 117  99 105 100  95  97 117  99  96  99 103  97 103  99  98  93
#>  [91]  91 100 103 105 101 118  87 110 111 119
#> 
#> $N_treatment
#>   [1]  98 100  96 100 100  99  97  82  74 100  93  96  90  84 106  78  99  94
#>  [19]  97  98 103  92 106  92  98 111  92  88  84 109  90  93 104 101 111 109
#>  [37] 108 107 103 105  93  99 102  94 102 107  82 100  97  95  98  99 113 103
#>  [55]  98 101 100 104  84 104  97  98  96  86  92  82 106  94  96  94 104 109
#>  [73]  97 102  83 101  95 100 105 103  83 101 104 101  97 103  97 101 102 107
#>  [91] 109 100  97  95  99  82 113  90  89  81
#> 
#> $early_stop
#>   [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [38] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [75] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $randomization_ratio
#>        [,1]      [,2]
#>   [1,]  0.5 0.4807692
#>   [2,]  0.5 0.4652406
#>   [3,]  0.5 0.4561957
#>   [4,]  0.5 0.4768831
#>   [5,]  0.5 0.4303165
#>   [6,]  0.5 0.4848485
#>   [7,]  0.5 0.4447537
#>   [8,]  0.5 0.4617176
#>   [9,]  0.5 0.3540773
#>  [10,]  0.5 0.4978723
#>  [11,]  0.5 0.5303030
#>  [12,]  0.5 0.4499222
#>  [13,]  0.5 0.4461538
#>  [14,]  0.5 0.4323034
#>  [15,]  0.5 0.4691943
#>  [16,]  0.5 0.4489796
#>  [17,]  0.5 0.4923599
#>  [18,]  0.5 0.4957959
#>  [19,]  0.5 0.4776306
#>  [20,]  0.5 0.5103237
#>  [21,]  0.5 0.4900352
#>  [22,]  0.5 0.4925466
#>  [23,]  0.5 0.4805970
#>  [24,]  0.5 0.4753363
#>  [25,]  0.5 0.4392971
#>  [26,]  0.5 0.5333333
#>  [27,]  0.5 0.4857143
#>  [28,]  0.5 0.4557486
#>  [29,]  0.5 0.4629393
#>  [30,]  0.5 0.4794441
#>  [31,]  0.5 0.5069930
#>  [32,]  0.5 0.3887868
#>  [33,]  0.5 0.5106762
#>  [34,]  0.5 0.5096154
#>  [35,]  0.5 0.5145985
#>  [36,]  0.5 0.4889019
#>  [37,]  0.5 0.4915068
#>  [38,]  0.5 0.4594735
#>  [39,]  0.5 0.4588501
#>  [40,]  0.5 0.4686181
#>  [41,]  0.5 0.4472033
#>  [42,]  0.5 0.4862814
#>  [43,]  0.5 0.4568672
#>  [44,]  0.5 0.5178781
#>  [45,]  0.5 0.4588104
#>  [46,]  0.5 0.5203046
#>  [47,]  0.5 0.4536501
#>  [48,]  0.5 0.5103237
#>  [49,]  0.5 0.5153566
#>  [50,]  0.5 0.4779858
#>  [51,]  0.5 0.4561957
#>  [52,]  0.5 0.4632510
#>  [53,]  0.5 0.4939759
#>  [54,]  0.5 0.4735119
#>  [55,]  0.5 0.4981009
#>  [56,]  0.5 0.5161290
#>  [57,]  0.5 0.5031056
#>  [58,]  0.5 0.4985809
#>  [59,]  0.5 0.4936526
#>  [60,]  0.5 0.4989282
#>  [61,]  0.5 0.5074167
#>  [62,]  0.5 0.4723336
#>  [63,]  0.5 0.4924044
#>  [64,]  0.5 0.4748437
#>  [65,]  0.5 0.4896074
#>  [66,]  0.5 0.4550708
#>  [67,]  0.5 0.4954700
#>  [68,]  0.5 0.4476302
#>  [69,]  0.5 0.4645309
#>  [70,]  0.5 0.4668588
#>  [71,]  0.5 0.4925723
#>  [72,]  0.5 0.5857143
#>  [73,]  0.5 0.4728493
#>  [74,]  0.5 0.4775918
#>  [75,]  0.5 0.3932584
#>  [76,]  0.5 0.4529412
#>  [77,]  0.5 0.4978225
#>  [78,]  0.5 0.4593274
#>  [79,]  0.5 0.4705882
#>  [80,]  0.5 0.4667089
#>  [81,]  0.5 0.4246575
#>  [82,]  0.5 0.4673367
#>  [83,]  0.5 0.5170486
#>  [84,]  0.5 0.4602442
#>  [85,]  0.5 0.4807122
#>  [86,]  0.5 0.4911334
#>  [87,]  0.5 0.4906388
#>  [88,]  0.5 0.5139241
#>  [89,]  0.5 0.4891138
#>  [90,]  0.5 0.5057724
#>  [91,]  0.5 0.5367447
#>  [92,]  0.5 0.5075956
#>  [93,]  0.5 0.5353132
#>  [94,]  0.5 0.4774920
#>  [95,]  0.5 0.4825702
#>  [96,]  0.5 0.4260870
#>  [97,]  0.5 0.4921956
#>  [98,]  0.5 0.4266319
#>  [99,]  0.5 0.4580316
#> [100,]  0.5 0.4635514
#> 

binomialfreq(p_control = 0.5, p_treatment = 0.40, N_total = 200,
            block_number = 2, simulation = 100, drift = -0.15)

#> $power
#> [1] 0
#> 
#> $p_control_estimate
#>   [1] 0.4343434 0.4344262 0.4818182 0.3775510 0.4272727 0.3760684 0.4660194
#>   [8] 0.4444444 0.3796296 0.4313725 0.3773585 0.4077670 0.3939394 0.3762376
#>  [15] 0.4000000 0.3750000 0.4056604 0.4827586 0.4356436 0.3928571 0.4901961
#>  [22] 0.4230769 0.4220183 0.4509804 0.3513514 0.4814815 0.3813559 0.3928571
#>  [29] 0.4838710 0.4596774 0.4431818 0.4183673 0.3750000 0.4000000 0.3655914
#>  [36] 0.4414414 0.4021739 0.3853211 0.4130435 0.2826087 0.3983051 0.4311927
#>  [43] 0.4793388 0.4141414 0.4301075 0.4380952 0.4466019 0.5045045 0.3763441
#>  [50] 0.4426230 0.3655914 0.4716981 0.4196429 0.4693878 0.3925234 0.4086022
#>  [57] 0.4571429 0.3867925 0.4333333 0.4955752 0.4827586 0.3627451 0.3684211
#>  [64] 0.3627451 0.3846154 0.5000000 0.4390244 0.4536082 0.4017857 0.3928571
#>  [71] 0.3693694 0.5673077 0.4333333 0.4056604 0.4324324 0.3461538 0.3482143
#>  [78] 0.4112150 0.3865546 0.4247788 0.3939394 0.3679245 0.4367816 0.4693878
#>  [85] 0.4387755 0.4285714 0.4901961 0.4196429 0.4905660 0.4629630 0.3750000
#>  [92] 0.3821138 0.4112150 0.3298969 0.3980583 0.3465347 0.5178571 0.4200000
#>  [99] 0.5233645 0.4642857
#> 
#> $p_treatment_estimate
#>   [1] 0.3267327 0.3076923 0.3111111 0.3725490 0.3222222 0.3253012 0.2783505
#>   [8] 0.2871287 0.3260870 0.3469388 0.2872340 0.3608247 0.4059406 0.2929293
#>  [15] 0.2900000 0.2980769 0.3297872 0.2619048 0.3838384 0.3522727 0.3367347
#>  [22] 0.3854167 0.3296703 0.3367347 0.3146067 0.3152174 0.2682927 0.3793103
#>  [29] 0.4392523 0.2368421 0.4107143 0.3627451 0.3839286 0.3176471 0.3644860
#>  [36] 0.3595506 0.3611111 0.3296703 0.3425926 0.3518519 0.2195122 0.3406593
#>  [43] 0.4177215 0.3366337 0.3457944 0.2631579 0.3402062 0.2696629 0.4112150
#>  [50] 0.3205128 0.3831776 0.2553191 0.2500000 0.3921569 0.3333333 0.2897196
#>  [57] 0.4526316 0.2978723 0.3500000 0.2758621 0.3214286 0.3265306 0.3619048
#>  [64] 0.3571429 0.2812500 0.3571429 0.3474576 0.3203883 0.3181818 0.3409091
#>  [71] 0.2808989 0.3125000 0.2818182 0.3723404 0.3483146 0.2708333 0.2500000
#>  [78] 0.3440860 0.3209877 0.3563218 0.3663366 0.3404255 0.2831858 0.3431373
#>  [85] 0.2745098 0.3368421 0.3163265 0.3409091 0.2978723 0.3913043 0.2500000
#>  [92] 0.2987013 0.3010753 0.3300971 0.2989691 0.3434343 0.2954545 0.3500000
#>  [99] 0.3655914 0.3522727
#> 
#> $prop_diff_estimate
#>   [1] -0.1111561045 -0.1278969957 -0.1900247321 -0.0166324435 -0.1081632653
#>   [6] -0.0589939971 -0.1977469322 -0.1685823755 -0.0604575163 -0.0785340314
#>  [11] -0.0890631283 -0.0472851132  0.0054666937 -0.0860236615 -0.1146126054
#>  [16] -0.0838815789 -0.0735826297 -0.2470489039 -0.0572572573 -0.0395858709
#>  [21] -0.1534613846 -0.0430999595 -0.0956433637 -0.1157873540 -0.0461851475
#>  [26] -0.1882956879 -0.1597318539 -0.0006284039 -0.0429298067 -0.2349290780
#>  [31] -0.0256252563 -0.1021505376  0.0164744646 -0.0974067046 -0.0006030151
#>  [36] -0.0795244927 -0.0413290113 -0.0290901589 -0.0694724124  0.0780805205
#>  [41] -0.1782464847 -0.0937942187 -0.0725891538 -0.0792059354 -0.0976602238
#>  [46] -0.1872323068 -0.1161950632 -0.2431332655  0.0347738693 -0.1221101303
#>  [51]  0.0227318447 -0.2206001622 -0.1789340102 -0.0873942811 -0.0590954774
#>  [56] -0.1237029502 -0.0236506974 -0.0946817083 -0.1304347826 -0.2381538462
#>  [61] -0.1690406368 -0.0485300040  0.0056829714 -0.0104208417 -0.0903225806
#>  [66] -0.1471074380 -0.0879917184 -0.1480789200 -0.0977859779 -0.0519480519
#>  [71] -0.0890868597 -0.2695121951 -0.1378609191 -0.0341365462 -0.0866734486
#>  [76] -0.0761523046 -0.1256345178 -0.0696037015 -0.0639473125 -0.0817713697
#>  [81] -0.0272054411 -0.0239244069 -0.1530010173 -0.1343133306 -0.1641282565
#>  [86] -0.0999395527 -0.1902040816 -0.0787337662 -0.2068126521 -0.0694949495
#>  [91] -0.1211878010 -0.0933132660 -0.1126533897 -0.0022039671 -0.0981693824
#>  [96] -0.0058058058 -0.2154471545 -0.0761217949 -0.1658147618 -0.1158536585
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1]  99 122 110  98 110 117 103  99 108 102 106 103  99 101 100  96 106 116
#>  [19] 101 112 102 104 109 102 111 108 118  84  93 124  88  98  88 115  93 111
#>  [37]  92 109  92  92 118 109 121  99  93 105 103 111  93 122  93 106 112  98
#>  [55] 107  93 105 106 120 113 116 102  95 102 104 116  82  97 112 112 111 104
#>  [73]  90 106 111 104 112 107 119 113  99 106  87  98  98 105 102 112 106 108
#>  [91]  96 123 107  97 103 101 112 100 107 112
#> 
#> $N_treatment
#>   [1] 101  78  90 102  90  83  97 101  92  98  94  97 101  99 100 104  94  84
#>  [19]  99  88  98  96  91  98  89  92  82 116 107  76 112 102 112  85 107  89
#>  [37] 108  91 108 108  82  91  79 101 107  95  97  89 107  78 107  94  88 102
#>  [55]  93 107  95  94  80  87  84  98 105  98  96  84 118 103  88  88  89  96
#>  [73] 110  94  89  96  88  93  81  87 101  94 113 102 102  95  98  88  94  92
#>  [91] 104  77  93 103  97  99  88 100  93  88
#> 
#> $early_stop
#>   [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [38] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [75] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $randomization_ratio
#>        [,1]      [,2]
#>   [1,]  0.5 0.4543744
#>   [2,]  0.5 0.3858268
#>   [3,]  0.5 0.4186603
#>   [4,]  0.5 0.4710178
#>   [5,]  0.5 0.4878049
#>   [6,]  0.5 0.4422638
#>   [7,]  0.5 0.4159734
#>   [8,]  0.5 0.4295039
#>   [9,]  0.5 0.4807692
#>  [10,]  0.5 0.5370450
#>  [11,]  0.5 0.4211124
#>  [12,]  0.5 0.4534433
#>  [13,]  0.5 0.4978355
#>  [14,]  0.5 0.4807692
#>  [15,]  0.5 0.4216061
#>  [16,]  0.5 0.4169279
#>  [17,]  0.5 0.4519590
#>  [18,]  0.5 0.4042553
#>  [19,]  0.5 0.4519941
#>  [20,]  0.5 0.4214286
#>  [21,]  0.5 0.4725392
#>  [22,]  0.5 0.4600939
#>  [23,]  0.5 0.4140625
#>  [24,]  0.5 0.4711538
#>  [25,]  0.5 0.4475138
#>  [26,]  0.5 0.4206009
#>  [27,]  0.5 0.3139535
#>  [28,]  0.5 0.5237484
#>  [29,]  0.5 0.4876102
#>  [30,]  0.5 0.3164062
#>  [31,]  0.5 0.5321027
#>  [32,]  0.5 0.4129643
#>  [33,]  0.5 0.5813953
#>  [34,]  0.5 0.3719234
#>  [35,]  0.5 0.5323571
#>  [36,]  0.5 0.3902439
#>  [37,]  0.5 0.4750000
#>  [38,]  0.5 0.4285714
#>  [39,]  0.5 0.4949495
#>  [40,]  0.5 0.6144153
#>  [41,]  0.5 0.3621495
#>  [42,]  0.5 0.4278169
#>  [43,]  0.5 0.4915254
#>  [44,]  0.5 0.4108058
#>  [45,]  0.5 0.4134615
#>  [46,]  0.5 0.4253731
#>  [47,]  0.5 0.4414753
#>  [48,]  0.5 0.4013212
#>  [49,]  0.5 0.4590164
#>  [50,]  0.5 0.4344828
#>  [51,]  0.5 0.5016051
#>  [52,]  0.5 0.3816794
#>  [53,]  0.5 0.3428571
#>  [54,]  0.5 0.4422281
#>  [55,]  0.5 0.4480000
#>  [56,]  0.5 0.4134615
#>  [57,]  0.5 0.4807692
#>  [58,]  0.5 0.3902439
#>  [59,]  0.5 0.3846154
#>  [60,]  0.5 0.3907083
#>  [61,]  0.5 0.3868613
#>  [62,]  0.5 0.5166960
#>  [63,]  0.5 0.5288462
#>  [64,]  0.5 0.4676856
#>  [65,]  0.5 0.4660714
#>  [66,]  0.5 0.4332248
#>  [67,]  0.5 0.5335968
#>  [68,]  0.5 0.4511692
#>  [69,]  0.5 0.4092664
#>  [70,]  0.5 0.4211036
#>  [71,]  0.5 0.4416322
#>  [72,]  0.5 0.3750000
#>  [73,]  0.5 0.4903846
#>  [74,]  0.5 0.5308311
#>  [75,]  0.5 0.3902439
#>  [76,]  0.5 0.4380165
#>  [77,]  0.5 0.4096386
#>  [78,]  0.5 0.5192308
#>  [79,]  0.5 0.3690852
#>  [80,]  0.5 0.4349610
#>  [81,]  0.5 0.5476190
#>  [82,]  0.5 0.5004621
#>  [83,]  0.5 0.4618016
#>  [84,]  0.5 0.4414047
#>  [85,]  0.5 0.4568966
#>  [86,]  0.5 0.4686717
#>  [87,]  0.5 0.4110479
#>  [88,]  0.5 0.3692458
#>  [89,]  0.5 0.3571429
#>  [90,]  0.5 0.4513671
#>  [91,]  0.5 0.4736842
#>  [92,]  0.5 0.3950495
#>  [93,]  0.5 0.3750000
#>  [94,]  0.5 0.4583949
#>  [95,]  0.5 0.4255319
#>  [96,]  0.5 0.4781997
#>  [97,]  0.5 0.4122137
#>  [98,]  0.5 0.4807692
#>  [99,]  0.5 0.3805825
#> [100,]  0.5 0.4332248
#>