Block Design for Response-Adaptive Randomization for Binomial Data

Simulation for binomial counts for block design for response-adaptive randomization with time as a confounding

binomialfreq(p_control, p_treatment, N_total, block_number = 4,
  drift = 0, simulation = 10000, zvalue = c(1, 1.5, 2),
  rand_ratio = c(1, 1.5, 2, 2.5), conf_int = 0.95,
  alternative = "less", correct = FALSE, replace = TRUE,
  early_stop = FALSE)

Arguments

p_control
scalar. Proportion of events under the control arm.

p_treatment
scalar. Proportion of events under the treatment arm.

N_total
scalar. Total sample size.

block_number
scalar. Number of blocks or time levels. The default is set to 4. If block_number is set to 1. This is a traditional RCT design.

drift
scalar. The increase or decrease in proportion of event over time. In this case, the proportion of failure changes in each block by the number of patient accured over the total sample size. The full drift effect is seen in the final block.

simulation
scalar. Number of simulation to be ran. The default is set to 10000.

zvalue
vector. The z-value cutoff for corrected chi-square test statistics.

rand_ratio
vector. The randomization ratio is set based on the corrected chi-square test statistic. The length of rand_ratio should be the same length of zvalue.

conf_int
scalar. Confidence level of the interval.

alternative
character. A string specifying the alternative hypothesis, must be one of "less" (default) or "greater".

correct
logical. A logical indicating whether to apply continuity correction when computing the test statistic: one half is subtracted from all |O - E| differences; however, the correction will not be bigger than the differences themselves.

replace
logical. should sampling be with replacement? If replace is set to FALSE (default), the 0 for control, 1 for treatment is replicated to the closest integer and this vector is sampled with no replacement. If replace is set to TRUE, the sampling is done based on randomization ratio provided with replacement.

early_stop
logical. A logical indicating whether the trials are stopped early for success or futility.

Value

a list with details on the simulation.

power

scalar. The power of the trial, ie. the proportion of success over the number of simulation ran.

p_control_estimate

scalar. The estimated proportion of events under the control group.

p_treatment_estimate

scalar. The estimated proportion of events under the treatment group.

N_enrolled

vector. The number of patients enrolled in the trial (sum of control and experimental group for each simulation. )

N_control

vector. The number of patients enrolled in the control group for each simulation.

N_control

vector. The number of patients enrolled in the experimental group for each simulation.

Examples

binomialfreq(p_control = 0.7, p_treatment = 0.65, N_total = 200,
            block_number = 2, simulation = 100)

#> $power
#> [1] 0.23
#> 
#> $p_control_estimate
#>   [1] 0.7291667 0.7142857 0.7362637 0.6526316 0.6842105 0.7087379 0.7096774
#>   [8] 0.7755102 0.7307692 0.6774194 0.7628866 0.7065217 0.6511628 0.7019231
#>  [15] 0.6237624 0.7692308 0.7745098 0.6565657 0.7894737 0.7040816 0.7641509
#>  [22] 0.6700000 0.7311828 0.6637168 0.7525773 0.7000000 0.6442308 0.6813187
#>  [29] 0.6483516 0.5934066 0.7090909 0.6730769 0.7706422 0.7722772 0.7767857
#>  [36] 0.6972477 0.6952381 0.6635514 0.7127660 0.7204301 0.6428571 0.7196262
#>  [43] 0.7238095 0.6666667 0.6728972 0.6605505 0.6542056 0.7222222 0.7100000
#>  [50] 0.7471264 0.6960784 0.6990291 0.6285714 0.7196262 0.6896552 0.6981132
#>  [57] 0.7113402 0.6500000 0.6000000 0.6213592 0.6631579 0.6634615 0.7472527
#>  [64] 0.6744186 0.6310680 0.6800000 0.7083333 0.6039604 0.6203704 0.6320755
#>  [71] 0.7319588 0.6736842 0.7247706 0.6853933 0.7659574 0.6666667 0.7037037
#>  [78] 0.6800000 0.6804124 0.6597938 0.6893204 0.5784314 0.7524752 0.7476636
#>  [85] 0.6262626 0.7363636 0.7623762 0.7596154 0.7391304 0.6666667 0.6930693
#>  [92] 0.6904762 0.7300000 0.6526316 0.6867470 0.7604167 0.6451613 0.6857143
#>  [99] 0.7173913 0.6767677
#> 
#> $p_treatment_estimate
#>   [1] 0.5576923 0.5504587 0.6055046 0.6571429 0.6627907 0.6804124 0.5514019
#>   [8] 0.5882353 0.5833333 0.6074766 0.6116505 0.6481481 0.6140351 0.6875000
#>  [15] 0.7474747 0.5729167 0.6734694 0.6831683 0.7333333 0.6666667 0.6702128
#>  [22] 0.6700000 0.7102804 0.7241379 0.5922330 0.6700000 0.6666667 0.6880734
#>  [29] 0.5871560 0.6330275 0.7222222 0.6875000 0.5934066 0.6161616 0.7159091
#>  [36] 0.5934066 0.6105263 0.6666667 0.6320755 0.7383178 0.6568627 0.6774194
#>  [43] 0.6105263 0.6734694 0.6774194 0.6153846 0.6129032 0.6956522 0.6800000
#>  [50] 0.6283186 0.6734694 0.5773196 0.6526316 0.7741935 0.6991150 0.6382979
#>  [57] 0.6601942 0.6300000 0.6111111 0.6701031 0.6857143 0.7187500 0.6055046
#>  [64] 0.6842105 0.6494845 0.7300000 0.6346154 0.6868687 0.6304348 0.6914894
#>  [71] 0.6796117 0.6761905 0.6043956 0.5765766 0.6981132 0.6506024 0.5869565
#>  [78] 0.6700000 0.5339806 0.6893204 0.6185567 0.6428571 0.5353535 0.6666667
#>  [85] 0.6435644 0.6222222 0.5959596 0.6458333 0.6944444 0.5981308 0.6363636
#>  [92] 0.6637931 0.6600000 0.6476190 0.5811966 0.6730769 0.6822430 0.5157895
#>  [99] 0.6666667 0.7227723
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1]  96  91  91  95 114 103  93  98 104  93  97  92  86 104 101 104 102  99
#>  [19]  95  98 106 100  93 113  97 100 104  91  91  91 110 104 109 101 112 109
#>  [37] 105 107  94  93  98 107 105 102 107 109 107 108 100  87 102 103 105 107
#>  [55]  87 106  97 100 110 103  95 104  91  86 103 100  96 101 108 106  97  95
#>  [73] 109  89  94 117 108 100  97  97 103 102 101 107  99 110 101 104  92  93
#>  [91] 101  84 100  95  83  96  93 105  92  99
#> 
#> $N_treatment
#>   [1] 104 109 109 105  86  97 107 102  96 107 103 108 114  96  99  96  98 101
#>  [19] 105 102  94 100 107  87 103 100  96 109 109 109  90  96  91  99  88  91
#>  [37]  95  93 106 107 102  93  95  98  93  91  93  92 100 113  98  97  95  93
#>  [55] 113  94 103 100  90  97 105  96 109 114  97 100 104  99  92  94 103 105
#>  [73]  91 111 106  83  92 100 103 103  97  98  99  93 101  90  99  96 108 107
#>  [91]  99 116 100 105 117 104 107  95 108 101
#> 

binomialfreq(p_control = 0.5, p_treatment = 0.40, N_total = 200,
            block_number = 2, simulation = 100, drift = -0.15)

#> $power
#> [1] 0.44
#> 
#> $p_control_estimate
#>   [1] 0.3846154 0.4059406 0.3925234 0.4137931 0.3981481 0.3831776 0.3406593
#>   [8] 0.3775510 0.4285714 0.3978495 0.4021739 0.3490566 0.3736264 0.3333333
#>  [15] 0.4157303 0.4456522 0.3854167 0.4117647 0.4536082 0.4065934 0.4659091
#>  [22] 0.5000000 0.3636364 0.4271845 0.3298969 0.3510638 0.3761468 0.3473684
#>  [29] 0.3620690 0.4000000 0.3944954 0.3809524 0.3448276 0.3956044 0.3783784
#>  [36] 0.3831776 0.4534884 0.3367347 0.3465347 0.4646465 0.4065934 0.3979592
#>  [43] 0.3168317 0.3838384 0.4897959 0.2755102 0.3775510 0.3465347 0.4851485
#>  [50] 0.3904762 0.5060241 0.4329897 0.3333333 0.4433962 0.4166667 0.3750000
#>  [57] 0.3725490 0.4392523 0.3398058 0.4687500 0.3786408 0.3800000 0.4086022
#>  [64] 0.3854167 0.4716981 0.4615385 0.4285714 0.3863636 0.3277311 0.4090909
#>  [71] 0.4196429 0.3592233 0.3750000 0.4174757 0.3225806 0.4158416 0.4038462
#>  [78] 0.3168317 0.4400000 0.3495146 0.4257426 0.3367347 0.3773585 0.4591837
#>  [85] 0.3736264 0.3617021 0.2886598 0.4408602 0.3867925 0.4803922 0.4455446
#>  [92] 0.4100000 0.3814433 0.4040404 0.4220183 0.4464286 0.4205607 0.3516484
#>  [99] 0.3203883 0.3673469
#> 
#> $p_treatment_estimate
#>   [1] 0.3486239 0.3030303 0.3333333 0.2743363 0.2934783 0.3870968 0.3394495
#>   [8] 0.3431373 0.2352941 0.3364486 0.2777778 0.2872340 0.3119266 0.3263158
#>  [15] 0.3693694 0.2592593 0.2211538 0.2244898 0.3106796 0.3577982 0.3035714
#>  [22] 0.3020833 0.2178218 0.3608247 0.2427184 0.2735849 0.2747253 0.2952381
#>  [29] 0.2976190 0.3700000 0.2417582 0.3052632 0.3362832 0.2385321 0.2696629
#>  [36] 0.2258065 0.2719298 0.3235294 0.3232323 0.2871287 0.2935780 0.3333333
#>  [43] 0.2424242 0.1980198 0.2058824 0.3431373 0.2549020 0.3333333 0.3434343
#>  [50] 0.2631579 0.2478632 0.2815534 0.2818182 0.3510638 0.3478261 0.3625000
#>  [57] 0.3469388 0.2903226 0.3814433 0.3365385 0.3814433 0.2600000 0.2523364
#>  [64] 0.2019231 0.2978723 0.3020833 0.3431373 0.2767857 0.3703704 0.2857143
#>  [71] 0.3522727 0.2989691 0.3409091 0.2474227 0.2990654 0.3838384 0.3437500
#>  [78] 0.2727273 0.2900000 0.3195876 0.3030303 0.3333333 0.3085106 0.2647059
#>  [85] 0.2477064 0.3018868 0.2427184 0.2990654 0.2234043 0.3265306 0.3434343
#>  [92] 0.2900000 0.2233010 0.3564356 0.2967033 0.3522727 0.2903226 0.3394495
#>  [99] 0.2783505 0.2156863
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1]  91 101 107  87 108 107  91  98  98  93  92 106  91 105  89  92  96 102
#>  [19]  97  91  88 104  99 103  97  94 109  95 116 100 109 105  87  91 111 107
#>  [37]  86  98 101  99  91  98 101  99  98  98  98 101 101 105  83  97  90 106
#>  [55] 108 120 102 107 103  96 103 100  93  96 106 104  98  88 119  88 112 103
#>  [73] 112 103  93 101 104 101 100 103 101  98 106  98  91  94  97  93 106 102
#>  [91] 101 100  97  99 109 112 107  91 103  98
#> 
#> $N_treatment
#>   [1] 109  99  93 113  92  93 109 102 102 107 108  94 109  95 111 108 104  98
#>  [19] 103 109 112  96 101  97 103 106  91 105  84 100  91  95 113 109  89  93
#>  [37] 114 102  99 101 109 102  99 101 102 102 102  99  99  95 117 103 110  94
#>  [55]  92  80  98  93  97 104  97 100 107 104  94  96 102 112  81 112  88  97
#>  [73]  88  97 107  99  96  99 100  97  99 102  94 102 109 106 103 107  94  98
#>  [91]  99 100 103 101  91  88  93 109  97 102
#>