Block Design for Response-Adaptive Randomization for Binomial Data

Simulation for binomial counts for block design for response-adaptive randomization with time as a confounding

binomialfreq(p_control, p_treatment, N_total, block_number = 4,
  drift = 0, simulation = 10000, conf_int = 0.95,
  alternative = "greater", correct = FALSE, replace = TRUE,
  early_stop = FALSE)

Arguments

p_control
scalar. Proportion of events under the control arm.

p_treatment
scalar. Proportion of events under the treatment arm.

N_total
scalar. Total sample size.

block_number
scalar. Number of blocks or time levels. The default is set to 4. If block_number is set to 1. This is a traditional RCT design.

drift
scalar. The increase or decrease in proportion of event over time. In this case, the proportion of failure changes in each block by the number of patient accured over the total sample size. The full drift effect is seen in the final block.

simulation
scalar. Number of simulation to be ran. The default is set to 10000.

conf_int
scalar. Confidence level of the interval.

alternative
character. A string specifying the alternative hypothesis, must be one of "less" or "greater" (default).

correct
logical. A logical indicating whether to apply continuity correction when computing the test statistic: one half is subtracted from all |O - E| differences; however, the correction will not be bigger than the differences themselves.

replace
logical. should sampling be with replacement? If replace is set to FALSE (default), the 0 for control, 1 for treatment is replicated to the closest integer and this vector is sampled with no replacement. If replace is set to TRUE, the sampling is done based on randomization ratio provided with replacement.

early_stop
logical. A logical indicating whether the trials are stopped early for success or futility.

Value

a list with details on the simulation.

power

scalar. The power of the trial, ie. the proportion of success over the number of simulation ran.

p_control_estimate

scalar. The estimated proportion of events under the control group.

p_treatment_estimate

scalar. The estimated proportion of events under the treatment group.

N_enrolled

vector. The number of patients enrolled in the trial (sum of control and experimental group for each simulation. )

N_control

vector. The number of patients enrolled in the control group for each simulation.

N_control

vector. The number of patients enrolled in the experimental group for each simulation.

Examples

binomialfreq(p_control = 0.7, p_treatment = 0.65, N_total = 200,
            block_number = 2, simulation = 100)

#> $power
#> [1] 0.01
#> 
#> $p_control_estimate
#>   [1] 0.7425743 0.7100000 0.7156863 0.7700000 0.7525773 0.6831683 0.7272727
#>   [8] 0.6810345 0.7118644 0.7300000 0.6388889 0.8100000 0.6226415 0.7565217
#>  [15] 0.7391304 0.7068966 0.7272727 0.7075472 0.6930693 0.7058824 0.6979167
#>  [22] 0.6481481 0.6595745 0.6698113 0.7653061 0.6888889 0.7314815 0.7117117
#>  [29] 0.6956522 0.6813187 0.7363636 0.7572816 0.7373737 0.7000000 0.6521739
#>  [36] 0.6666667 0.7173913 0.6888889 0.6736842 0.7065217 0.7333333 0.6732673
#>  [43] 0.6770833 0.6542056 0.6907216 0.7127660 0.6991150 0.6800000 0.6893204
#>  [50] 0.7238095 0.7244898 0.7300000 0.6321839 0.6847826 0.7254902 0.6161616
#>  [57] 0.5900000 0.6041667 0.6810345 0.7291667 0.6826923 0.7373737 0.7500000
#>  [64] 0.7079646 0.7222222 0.7217391 0.6808511 0.6764706 0.6730769 0.6857143
#>  [71] 0.6666667 0.5625000 0.6568627 0.6836735 0.8018018 0.7113402 0.7428571
#>  [78] 0.7812500 0.6559140 0.6770833 0.7433628 0.6804124 0.6804124 0.7500000
#>  [85] 0.6893204 0.7525773 0.6893204 0.6237624 0.7244898 0.7311828 0.6595745
#>  [92] 0.7524752 0.6981132 0.7378641 0.6868687 0.7522124 0.7356322 0.6320755
#>  [99] 0.7407407 0.6637931
#> 
#> $p_treatment_estimate
#>   [1] 0.7171717 0.5900000 0.6326531 0.6200000 0.6116505 0.6060606 0.5544554
#>   [8] 0.7023810 0.5365854 0.6600000 0.7282609 0.7300000 0.5957447 0.6352941
#>  [15] 0.6018519 0.6785714 0.6039604 0.6914894 0.6666667 0.6632653 0.6250000
#>  [22] 0.6413043 0.6226415 0.5744681 0.6568627 0.6909091 0.6847826 0.6516854
#>  [29] 0.5764706 0.6238532 0.7555556 0.5979381 0.7029703 0.7100000 0.6574074
#>  [36] 0.6454545 0.7222222 0.7000000 0.6000000 0.5740741 0.6526316 0.5858586
#>  [43] 0.6153846 0.6881720 0.6990291 0.6603774 0.6551724 0.6100000 0.7319588
#>  [50] 0.6631579 0.6470588 0.6000000 0.5929204 0.6388889 0.6530612 0.6831683
#>  [57] 0.6600000 0.5961538 0.6666667 0.6826923 0.6979167 0.5544554 0.6562500
#>  [64] 0.6091954 0.6847826 0.6941176 0.6132075 0.5612245 0.6145833 0.6526316
#>  [71] 0.7211538 0.7884615 0.6530612 0.6666667 0.5955056 0.6310680 0.6736842
#>  [78] 0.6057692 0.6355140 0.6153846 0.6091954 0.6116505 0.6796117 0.6346154
#>  [85] 0.6494845 0.7669903 0.6804124 0.6464646 0.6274510 0.7383178 0.7169811
#>  [92] 0.6666667 0.6914894 0.6701031 0.6732673 0.6206897 0.7079646 0.5957447
#>  [99] 0.5978261 0.5952381
#> 
#> $prop_diff_estimate
#>   [1] -0.0284740325 -0.1217948718 -0.0859719439 -0.1510604242 -0.1407407407
#>   [6] -0.0770154031 -0.1796733212  0.0205338809 -0.1728913135 -0.0642594859
#>  [11]  0.0964343598 -0.0780312125 -0.0318291700 -0.1189743590 -0.1372785829
#>  [16] -0.0379510577 -0.1241098678 -0.0201045436 -0.0363819095 -0.0444579781
#>  [21] -0.0715447154 -0.0119967466 -0.0358581789 -0.1023321271 -0.1080864692
#>  [26]  0.0016168149 -0.0421393841 -0.0601336303 -0.1233806292 -0.0575176589
#>  [31]  0.0180423061 -0.1409081653 -0.0364948867  0.0090036014  0.0066451873
#>  [36] -0.0259686727  0.0072580645  0.0054900366 -0.0824030850 -0.1371717172
#>  [41] -0.0731756620 -0.0892462312 -0.0645161290  0.0321866828  0.0058197873
#>  [46] -0.0502008032 -0.0435401831 -0.0614035088  0.0432779002 -0.0647452811
#>  [51] -0.0774309724 -0.1238458450 -0.0432820513 -0.0422936153 -0.0724579664
#>  [56]  0.0660132026  0.0681089744 -0.0084168337 -0.0137577002 -0.0462925852
#>  [61]  0.0194253339 -0.1706823244 -0.0999182673 -0.0942547536 -0.0406060606
#>  [66] -0.0306395229 -0.0670682731 -0.1248459959 -0.0573836276 -0.0326849810
#>  [71]  0.0479562930  0.2263242376  0.0016064257  0.0010421009 -0.2048891651
#>  [76] -0.0737120749 -0.0715585203 -0.1712907117 -0.0183801252 -0.0532258065
#>  [81] -0.1369863014 -0.0690690691 -0.0006006006 -0.1158021713 -0.0406406406
#>  [86]  0.0162292126 -0.0090090090  0.0164824121 -0.1081353202  0.0116864800
#>  [91]  0.0593360996 -0.0846944949 -0.0066238458 -0.0678678679 -0.0112112112
#>  [96] -0.1323679380 -0.0297776871 -0.0922454197 -0.1331842212 -0.0690789474
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1] 101 100 102 100  97 101  99 116 118 100 108 100 106 115  92 116  99 106
#>  [19] 101 102  96 108  94 106  98  90 108 111 115  91 110 103  99 100  92  90
#>  [37]  92  90  95  92 105 101  96 107  97  94 113 100 103 105  98 100  87  92
#>  [55] 102  99 100  96 116  96 104  99 104 113 108 115  94 102 104 105  96  96
#>  [73] 102  98 111  97 105  96  93  96 113  97  97  96 103  97 103 101  98  93
#>  [91]  94 101 106 103  99 113  87 106 108 116
#> 
#> $N_treatment
#>   [1]  99 100  98 100 103  99 101  84  82 100  92 100  94  85 108  84 101  94
#>  [19]  99  98 104  92 106  94 102 110  92  89  85 109  90  97 101 100 108 110
#>  [37] 108 110 105 108  95  99 104  93 103 106  87 100  97  95 102 100 113 108
#>  [55]  98 101 100 104  84 104  96 101  96  87  92  85 106  98  96  95 104 104
#>  [73]  98 102  89 103  95 104 107 104  87 103 103 104  97 103  97  99 102 107
#>  [91] 106  99  94  97 101  87 113  94  92  84
#> 

binomialfreq(p_control = 0.5, p_treatment = 0.40, N_total = 200,
            block_number = 2, simulation = 100, drift = -0.15)

#> $power
#> [1] 0
#> 
#> $p_control_estimate
#>   [1] 0.3775510 0.4117647 0.4476190 0.3500000 0.3909091 0.3304348 0.4554455
#>   [8] 0.3645833 0.3703704 0.3689320 0.3714286 0.3700000 0.3366337 0.3366337
#>  [15] 0.3684211 0.3478261 0.3653846 0.4247788 0.4062500 0.3693694 0.4700000
#>  [22] 0.4038462 0.3853211 0.4158416 0.3272727 0.4476190 0.3577982 0.3837209
#>  [29] 0.4456522 0.4152542 0.4333333 0.3804348 0.3186813 0.3761468 0.3617021
#>  [36] 0.4128440 0.3655914 0.3653846 0.3736264 0.2772277 0.3846154 0.4000000
#>  [43] 0.4508197 0.3775510 0.4000000 0.3838384 0.4554455 0.4245283 0.3655914
#>  [50] 0.4000000 0.3548387 0.4200000 0.4019608 0.4315789 0.3644860 0.3777778
#>  [57] 0.4112150 0.3495146 0.4181818 0.4732143 0.4778761 0.3300971 0.3578947
#>  [64] 0.3431373 0.3365385 0.4782609 0.3571429 0.4270833 0.3944954 0.3423423
#>  [71] 0.3333333 0.5050505 0.2777778 0.3584906 0.4144144 0.3137255 0.2871287
#>  [78] 0.3831776 0.3363636 0.4234234 0.3400000 0.3750000 0.3953488 0.4270833
#>  [85] 0.4226804 0.3904762 0.4747475 0.3669725 0.4800000 0.4351852 0.3018868
#>  [92] 0.3781513 0.4174757 0.2842105 0.3673469 0.2970297 0.4403670 0.3800000
#>  [99] 0.5000000 0.4166667
#> 
#> $p_treatment_estimate
#>   [1] 0.2843137 0.2345679 0.2842105 0.3100000 0.3000000 0.2941176 0.2323232
#>   [8] 0.2788462 0.3043478 0.3298969 0.2526316 0.3300000 0.3535354 0.2424242
#>  [15] 0.2666667 0.2777778 0.3020833 0.2413793 0.3076923 0.3258427 0.3300000
#>  [22] 0.2916667 0.2967033 0.3232323 0.2555556 0.2631579 0.2417582 0.3421053
#>  [29] 0.3796296 0.2073171 0.3909091 0.2962963 0.3761468 0.2637363 0.3113208
#>  [36] 0.2747253 0.2803738 0.2812500 0.3211009 0.2828283 0.1686747 0.2842105
#>  [43] 0.3717949 0.2941176 0.3000000 0.2277228 0.2929293 0.2446809 0.3457944
#>  [50] 0.3000000 0.3551402 0.2400000 0.2040816 0.3428571 0.3225806 0.2545455
#>  [57] 0.3870968 0.2886598 0.2888889 0.2386364 0.2528736 0.2886598 0.3428571
#>  [64] 0.3163265 0.2291667 0.3411765 0.2931034 0.2596154 0.2417582 0.2921348
#>  [71] 0.2391304 0.2673267 0.3369565 0.3191489 0.2921348 0.2142857 0.2323232
#>  [78] 0.3010753 0.2777778 0.3370787 0.3400000 0.2812500 0.2456140 0.2980769
#>  [85] 0.2427184 0.2842105 0.3069307 0.2857143 0.2900000 0.3478261 0.3297872
#>  [92] 0.2592593 0.2680412 0.2761905 0.2843137 0.3030303 0.2747253 0.3200000
#>  [99] 0.3367347 0.3260870
#> 
#> $prop_diff_estimate
#>   [1] -0.094179466 -0.179449464 -0.173842733 -0.048780488 -0.091836735
#>   [6] -0.041025641 -0.230198516 -0.094491355 -0.074346405 -0.030778515
#>  [11] -0.118001204 -0.038940185  0.015869786 -0.099458592 -0.102666934
#>  [16] -0.075980392 -0.060514676 -0.204098530 -0.095591182 -0.041742655
#>  [21] -0.136054422 -0.117361392 -0.091185410 -0.094472362 -0.078669911
#>  [26] -0.200121778 -0.149520767 -0.032298137 -0.070760472 -0.218143460
#>  [31] -0.039446930 -0.118464052  0.063427526 -0.119412125 -0.050381373
#>  [36] -0.139255946 -0.087053121 -0.052170284 -0.050454087  0.006401280
#>  [41] -0.215213358 -0.115700822 -0.092608326 -0.084769539 -0.109597397
#>  [46] -0.161809045 -0.167767768 -0.183534137 -0.019698492 -0.097540642
#>  [51]  0.005431503 -0.181891026 -0.198868230 -0.095477387 -0.041608040
#>  [56] -0.130540870 -0.040081384 -0.063313965 -0.151020408 -0.249183007
#>  [61] -0.232337796 -0.054490414 -0.004059265 -0.031663327 -0.092741935
#>  [66] -0.139214477 -0.062012320 -0.187991718 -0.161296859 -0.050010124
#>  [71] -0.093435360 -0.245226131  0.044530297 -0.039357430 -0.125737538
#>  [76] -0.099439776 -0.067957249 -0.084892376 -0.061232766 -0.097162686
#>  [81]  0.000000000 -0.086851628 -0.149734802 -0.130679534 -0.178406582
#>  [86] -0.109611122 -0.183235473 -0.080121090 -0.194711538 -0.085858586
#>  [91]  0.024979855 -0.125549278 -0.148348348 -0.008421897 -0.083666934
#>  [96]  0.004204204 -0.155825735 -0.066506410 -0.169150222 -0.090579710
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1]  98 119 105 100 110 115 101  96 108 103 105 100 101 101  95  92 104 113
#>  [19]  96 111 100 104 109 101 110 105 109  86  92 118  90  92  91 109  94 109
#>  [37]  93 104  91 101 117 105 122  98  90  99 101 106  93 120  93 100 102  95
#>  [55] 107  90 107 103 110 112 113 103  95 102 104 115  84  96 109 111 108  99
#>  [73] 108 106 111 102 101 107 110 111 100 104  86  96  97 105  99 109 100 108
#>  [91] 106 119 103  95  98 101 109 100 102 108
#> 
#> $N_treatment
#>   [1] 102  81  95 100  90  85  99 104  92  97  95 100  99  99 105 108  96  87
#>  [19] 104  89 100  96  91  99  90  95  91 114 108  82 110 108 109  91 106  91
#>  [37] 107  96 109  99  83  95  78 102 110 101  99  94 107  80 107 100  98 105
#>  [55]  93 110  93  97  90  88  87  97 105  98  96  85 116 104  91  89  92 101
#>  [73]  92  94  89  98  99  93  90  89 100  96 114 104 103  95 101  91 100  92
#>  [91]  94  81  97 105 102  99  91 100  98  92
#>