Block Design for Response-Adaptive Randomization for Binomial Data

Simulation for binomial counts for block design for response-adaptive randomization with time as a confounding

binomialfreq(p_control, p_treatment, N_total, block_number = 4,
  drift = 0, simulation = 10000, conf_int = 0.95,
  alternative = "greater", correct = FALSE, replace = TRUE,
  early_stop = FALSE)

Arguments

p_control
scalar. Proportion of events under the control arm.

p_treatment
scalar. Proportion of events under the treatment arm.

N_total
scalar. Total sample size.

block_number
scalar. Number of blocks or time levels. The default is set to 4. If block_number is set to 1. This is a traditional RCT design.

drift
scalar. The increase or decrease in proportion of event over time. In this case, the proportion of failure changes in each block by the number of patient accured over the total sample size. The full drift effect is seen in the final block.

simulation
scalar. Number of simulation to be ran. The default is set to 10000.

conf_int
scalar. Confidence level of the interval.

alternative
character. A string specifying the alternative hypothesis, must be one of "less" or "greater" (default).

correct
logical. A logical indicating whether to apply continuity correction when computing the test statistic: one half is subtracted from all |O - E| differences; however, the correction will not be bigger than the differences themselves.

replace
logical. should sampling be with replacement? If replace is set to FALSE (default), the 0 for control, 1 for treatment is replicated to the closest integer and this vector is sampled with no replacement. If replace is set to TRUE, the sampling is done based on randomization ratio provided with replacement.

early_stop
logical. A logical indicating whether the trials are stopped early for success or futility.

Value

a list with details on the simulation.

power

scalar. The power of the trial, ie. the proportion of success over the number of simulation ran.

p_control_estimate

scalar. The estimated proportion of events under the control group.

p_treatment_estimate

scalar. The estimated proportion of events under the treatment group.

N_enrolled

vector. The number of patients enrolled in the trial (sum of control and experimental group for each simulation. )

N_control

vector. The number of patients enrolled in the control group for each simulation.

N_control

vector. The number of patients enrolled in the experimental group for each simulation.

Examples

binomialfreq(p_control = 0.7, p_treatment = 0.65, N_total = 200,
            block_number = 2, simulation = 100)

#> $power
#> [1] 0.01
#> 
#> $p_control_estimate
#>   [1] 0.7425743 0.7100000 0.7156863 0.7700000 0.7525773 0.6831683 0.7272727
#>   [8] 0.6810345 0.7118644 0.7300000 0.6388889 0.8100000 0.6226415 0.7565217
#>  [15] 0.7391304 0.7068966 0.7272727 0.7075472 0.6930693 0.7058824 0.6979167
#>  [22] 0.6481481 0.6595745 0.6698113 0.7653061 0.6888889 0.7314815 0.7117117
#>  [29] 0.6956522 0.6813187 0.7363636 0.7572816 0.7373737 0.7000000 0.6521739
#>  [36] 0.6666667 0.7173913 0.6888889 0.6736842 0.7065217 0.7333333 0.6732673
#>  [43] 0.6770833 0.6542056 0.6907216 0.7127660 0.6991150 0.6800000 0.6893204
#>  [50] 0.7238095 0.7244898 0.7300000 0.6321839 0.6847826 0.7254902 0.6161616
#>  [57] 0.5900000 0.6041667 0.6810345 0.7291667 0.6826923 0.7373737 0.7500000
#>  [64] 0.7079646 0.7222222 0.7217391 0.6808511 0.6764706 0.6730769 0.6857143
#>  [71] 0.6666667 0.5625000 0.6568627 0.6836735 0.8018018 0.7113402 0.7428571
#>  [78] 0.7812500 0.6559140 0.6770833 0.7433628 0.6804124 0.6804124 0.7500000
#>  [85] 0.6893204 0.7525773 0.6893204 0.6237624 0.7244898 0.7311828 0.6595745
#>  [92] 0.7524752 0.6981132 0.7378641 0.6868687 0.7522124 0.7356322 0.6320755
#>  [99] 0.7407407 0.6637931
#> 
#> $p_treatment_estimate
#>   [1] 0.7171717 0.5900000 0.6326531 0.6200000 0.6116505 0.6060606 0.5544554
#>   [8] 0.7023810 0.5365854 0.6600000 0.7282609 0.7300000 0.5957447 0.6352941
#>  [15] 0.6018519 0.6785714 0.6039604 0.6914894 0.6666667 0.6632653 0.6250000
#>  [22] 0.6413043 0.6226415 0.5744681 0.6568627 0.6909091 0.6847826 0.6516854
#>  [29] 0.5764706 0.6238532 0.7555556 0.5979381 0.7029703 0.7100000 0.6574074
#>  [36] 0.6454545 0.7222222 0.7000000 0.6000000 0.5740741 0.6526316 0.5858586
#>  [43] 0.6153846 0.6881720 0.6990291 0.6603774 0.6551724 0.6100000 0.7319588
#>  [50] 0.6631579 0.6470588 0.6000000 0.5929204 0.6388889 0.6530612 0.6831683
#>  [57] 0.6600000 0.5961538 0.6666667 0.6826923 0.6979167 0.5544554 0.6562500
#>  [64] 0.6091954 0.6847826 0.6941176 0.6132075 0.5612245 0.6145833 0.6526316
#>  [71] 0.7211538 0.7884615 0.6530612 0.6666667 0.5955056 0.6310680 0.6736842
#>  [78] 0.6057692 0.6355140 0.6153846 0.6091954 0.6116505 0.6796117 0.6346154
#>  [85] 0.6494845 0.7669903 0.6804124 0.6464646 0.6274510 0.7383178 0.7169811
#>  [92] 0.6666667 0.6914894 0.6701031 0.6732673 0.6206897 0.7079646 0.5957447
#>  [99] 0.5978261 0.5952381
#> 
#> $prop_diff_estimate
#>   [1] -0.0284740325 -0.1217948718 -0.0859719439 -0.1510604242 -0.1407407407
#>   [6] -0.0770154031 -0.1796733212  0.0205338809 -0.1728913135 -0.0642594859
#>  [11]  0.0964343598 -0.0780312125 -0.0318291700 -0.1189743590 -0.1372785829
#>  [16] -0.0379510577 -0.1241098678 -0.0201045436 -0.0363819095 -0.0444579781
#>  [21] -0.0715447154 -0.0119967466 -0.0358581789 -0.1023321271 -0.1080864692
#>  [26]  0.0016168149 -0.0421393841 -0.0601336303 -0.1233806292 -0.0575176589
#>  [31]  0.0180423061 -0.1409081653 -0.0364948867  0.0090036014  0.0066451873
#>  [36] -0.0259686727  0.0072580645  0.0054900366 -0.0824030850 -0.1371717172
#>  [41] -0.0731756620 -0.0892462312 -0.0645161290  0.0321866828  0.0058197873
#>  [46] -0.0502008032 -0.0435401831 -0.0614035088  0.0432779002 -0.0647452811
#>  [51] -0.0774309724 -0.1238458450 -0.0432820513 -0.0422936153 -0.0724579664
#>  [56]  0.0660132026  0.0681089744 -0.0084168337 -0.0137577002 -0.0462925852
#>  [61]  0.0194253339 -0.1706823244 -0.0999182673 -0.0942547536 -0.0406060606
#>  [66] -0.0306395229 -0.0670682731 -0.1248459959 -0.0573836276 -0.0326849810
#>  [71]  0.0479562930  0.2263242376  0.0016064257  0.0010421009 -0.2048891651
#>  [76] -0.0737120749 -0.0715585203 -0.1712907117 -0.0183801252 -0.0532258065
#>  [81] -0.1369863014 -0.0690690691 -0.0006006006 -0.1158021713 -0.0406406406
#>  [86]  0.0162292126 -0.0090090090  0.0164824121 -0.1081353202  0.0116864800
#>  [91]  0.0593360996 -0.0846944949 -0.0066238458 -0.0678678679 -0.0112112112
#>  [96] -0.1323679380 -0.0297776871 -0.0922454197 -0.1331842212 -0.0690789474
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1] 101 100 102 100  97 101  99 116 118 100 108 100 106 115  92 116  99 106
#>  [19] 101 102  96 108  94 106  98  90 108 111 115  91 110 103  99 100  92  90
#>  [37]  92  90  95  92 105 101  96 107  97  94 113 100 103 105  98 100  87  92
#>  [55] 102  99 100  96 116  96 104  99 104 113 108 115  94 102 104 105  96  96
#>  [73] 102  98 111  97 105  96  93  96 113  97  97  96 103  97 103 101  98  93
#>  [91]  94 101 106 103  99 113  87 106 108 116
#> 
#> $N_treatment
#>   [1]  99 100  98 100 103  99 101  84  82 100  92 100  94  85 108  84 101  94
#>  [19]  99  98 104  92 106  94 102 110  92  89  85 109  90  97 101 100 108 110
#>  [37] 108 110 105 108  95  99 104  93 103 106  87 100  97  95 102 100 113 108
#>  [55]  98 101 100 104  84 104  96 101  96  87  92  85 106  98  96  95 104 104
#>  [73]  98 102  89 103  95 104 107 104  87 103 103 104  97 103  97  99 102 107
#>  [91] 106  99  94  97 101  87 113  94  92  84
#> 

binomialfreq(p_control = 0.5, p_treatment = 0.40, N_total = 200,
            block_number = 2, simulation = 100, drift = -0.15)

#> $power
#> [1] 0
#> 
#> $p_control_estimate
#>   [1] 0.4387755 0.4375000 0.5000000 0.3854167 0.4272727 0.3706897 0.4752475
#>   [8] 0.4432990 0.3796296 0.4313725 0.3714286 0.4040404 0.3939394 0.3762376
#>  [15] 0.3804348 0.3695652 0.4038462 0.4912281 0.4489796 0.3909091 0.4851485
#>  [22] 0.4271845 0.4150943 0.4455446 0.3486239 0.4903846 0.3944954 0.3953488
#>  [29] 0.4891304 0.4661017 0.4382022 0.4361702 0.3763441 0.4112150 0.3723404
#>  [36] 0.4299065 0.3956044 0.3942308 0.4065934 0.2970297 0.4070796 0.4466019
#>  [43] 0.4833333 0.4081633 0.4252874 0.4326923 0.4554455 0.4857143 0.3763441
#>  [50] 0.4333333 0.3655914 0.4700000 0.4380952 0.4736842 0.3925234 0.4157303
#>  [57] 0.4615385 0.3786408 0.4454545 0.5000000 0.4867257 0.3653846 0.3684211
#>  [64] 0.3627451 0.3846154 0.5043478 0.4285714 0.4526316 0.4074074 0.3853211
#>  [71] 0.3796296 0.5656566 0.4333333 0.4056604 0.4363636 0.3529412 0.3495146
#>  [78] 0.4112150 0.3909091 0.4285714 0.3900000 0.3679245 0.4285714 0.4623656
#>  [85] 0.4479167 0.4285714 0.4949495 0.4311927 0.5000000 0.4716981 0.3789474
#>  [92] 0.3879310 0.4271845 0.3368421 0.3979592 0.3500000 0.5229358 0.4200000
#>  [99] 0.5196078 0.4636364
#> 
#> $p_treatment_estimate
#>   [1] 0.3235294 0.3068182 0.2978723 0.3653846 0.3222222 0.3333333 0.2727273
#>   [8] 0.2912621 0.3260870 0.3469388 0.2947368 0.3564356 0.4059406 0.2929293
#>  [15] 0.2962963 0.2870370 0.3333333 0.2558140 0.3725490 0.3555556 0.3434343
#>  [22] 0.3814433 0.3404255 0.3434343 0.3186813 0.3125000 0.2637363 0.3859649
#>  [29] 0.4351852 0.2439024 0.4144144 0.3490566 0.3831776 0.3010753 0.3584906
#>  [36] 0.3763441 0.3669725 0.3125000 0.3394495 0.3434343 0.2183908 0.3298969
#>  [43] 0.4125000 0.3333333 0.3539823 0.2604167 0.3333333 0.2842105 0.4112150
#>  [50] 0.3250000 0.3831776 0.2500000 0.2421053 0.3904762 0.3333333 0.2792793
#>  [57] 0.4479167 0.2989691 0.3444444 0.2727273 0.3218391 0.3229167 0.3619048
#>  [64] 0.3571429 0.2812500 0.3529412 0.3534483 0.3238095 0.3152174 0.3406593
#>  [71] 0.2717391 0.3168317 0.2818182 0.3723404 0.3444444 0.2653061 0.2474227
#>  [78] 0.3440860 0.3111111 0.3522727 0.3700000 0.3404255 0.2931034 0.3551402
#>  [85] 0.2692308 0.3368421 0.3168317 0.3296703 0.3000000 0.3829787 0.2476190
#>  [92] 0.2976190 0.2886598 0.3238095 0.2941176 0.3400000 0.2967033 0.3500000
#>  [99] 0.3775510 0.3555556
#> 
#> $prop_diff_estimate
#>   [1] -0.1184316896 -0.1308943089 -0.2153284672 -0.0304454434 -0.1081632653
#>   [6] -0.0442933663 -0.2095448165 -0.1605310803 -0.0604575163 -0.0785340314
#>  [11] -0.0758579169 -0.0478391960  0.0054666937 -0.0860236615 -0.0823600483
#>  [16] -0.0898692810 -0.0671491757 -0.2577019151 -0.0764305722 -0.0333872926
#>  [21] -0.1396279256 -0.0506559031 -0.0763972658 -0.1033165829 -0.0365878310
#>  [26] -0.1928369081 -0.1601703940  0.0004140787 -0.0524603497 -0.2282700422
#>  [31] -0.0175546030 -0.1232596233  0.0111088669 -0.1183383992 -0.0138498595
#>  [36] -0.0517067259 -0.0288676560 -0.0454924875 -0.0653884965  0.0472094419
#>  [41] -0.1874872214 -0.1161161161 -0.0805369128 -0.0765531062 -0.0770956557
#>  [46] -0.1845528455 -0.1279279279 -0.2027270904  0.0347738693 -0.1054606086
#>  [51]  0.0227318447 -0.2219551282 -0.2006934530 -0.0894472362 -0.0590954774
#>  [56] -0.1399796541 -0.0306451613 -0.0827489481 -0.1224489796 -0.2434640523
#>  [61] -0.1703398558 -0.0580736544  0.0056829714 -0.0104208417 -0.0903225806
#>  [66] -0.1550483241 -0.0733059548 -0.1414412554 -0.1004040404 -0.0431886983
#>  [71] -0.1067257350 -0.2570854271 -0.1378609191 -0.0341365462 -0.0940794809
#>  [76] -0.0876350540 -0.1167173252 -0.0696037015 -0.0794809408 -0.0883558836
#>  [81] -0.0200000000 -0.0239244069 -0.1367556468 -0.1100382217 -0.1787293928
#>  [86] -0.0999395527 -0.1903219275 -0.1013118063 -0.2051282051 -0.0872537193
#>  [91] -0.1266331658 -0.0943551710 -0.1377377377 -0.0138359735 -0.1036829464
#>  [96] -0.0118047219 -0.2145896657 -0.0761217949 -0.1461941200 -0.1099434115
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1]  98 112 106  96 110 116 101  97 108 102 105  99  99 101  92  92 104 114
#>  [19]  98 110 101 103 106 101 109 104 109  86  92 118  89  94  93 107  94 107
#>  [37]  91 104  91 101 113 103 120  98  87 104 101 105  93 120  93 100 105  95
#>  [55] 107  89 104 103 110 112 113 104  95 102 104 115  84  95 108 109 108  99
#>  [73]  90 106 110 102 103 107 110 112 100 106  84  93  96 105  99 109 100 106
#>  [91]  95 116 103  95  98 100 109 100 102 110
#> 
#> $N_treatment
#>   [1] 102  88  94 104  90  84  99 103  92  98  95 101 101  99 108 108  96  86
#>  [19] 102  90  99  97  94  99  91  96  91 114 108  82 111 106 107  93 106  93
#>  [37] 109  96 109  99  87  97  80 102 113  96  99  95 107  80 107 100  95 105
#>  [55]  93 111  96  97  90  88  87  96 105  98  96  85 116 105  92  91  92 101
#>  [73] 110  94  90  98  97  93  90  88 100  94 116 107 104  95 101  91 100  94
#>  [91] 105  84  97 105 102 100  91 100  98  90
#>