Block Design for Response-Adaptive Randomization for Binomial Data

Simulation for binomial counts for block design for response-adaptive randomization with time as a confounding

binomialfreq(p_control, p_treatment, N_total, block_number = 4,
  drift = 0, simulation = 10000, zvalue = c(1, 1.5, 2),
  rand_ratio = c(1, 1.5, 2, 2.5), conf_int = 0.95,
  alternative = "less", correct = FALSE, replace = TRUE)

Arguments

p_control
scalar. Proportion of events under the control arm.

p_treatment
scalar. Proportion of events under the treatment arm.

N_total
scalar. Total sample size.

block_number
scalar. Number of blocks or time levels. The default is set to 4. If block_number is set to 1. This is a traditional RCT design.

drift
scalar. The increase or decrease in proportion of event over time. In this case, the proportion of failure changes in each block by the number of patient accured over the total sample size. The full drift effect is seen in the final block.

simulation
scalar. Number of simulation to be ran. The default is set to 10000.

zvalue
vector. The z-value cutoff for corrected chi-square test statistics.

rand_ratio
vector. The randomization ratio is set based on the corrected chi-square test statistic. The length of rand_ratio should be the same length of zvalue.

conf_int
scalar. Confidence level of the interval.

alternative
character. A string specifying the alternative hypothesis, must be one of "less" (default) or "greater".

correct
logical. a logical indicating whether to apply continuity correction when computing the test statistic: one half is subtracted from all |O - E| differences; however, the correction will not be bigger than the differences themselves.

replace
logical. should sampling be with replacement? If replace is set to FALSE (default), the 0 for control, 1 for treatment is replicated to the closest integer and this vector is sampled with no replacement. If replace is set to TRUE, the sampling is done based on randomization ratio provided with replacement.

Value

a list with details on the simulation.

power

scalar. The power of the trial, ie. the proportion of success over the number of simulation ran.

p_control_estimate

scalar. The estimated proportion of events under the control group.

p_treatment_estimate

scalar. The estimated proportion of events under the treatment group.

N_enrolled

vector. The number of patients enrolled in the trial (sum of control and experimental group for each simulation. )

N_control

vector. The number of patients enrolled in the control group for each simulation.

N_control

vector. The number of patients enrolled in the experimental group for each simulation.

Examples

binomialfreq(p_control = 0.7, p_treatment = 0.65, N_total = 200,
            block_number = 2, simulation = 100)

#> $power
#> [1] 0.19
#> 
#> $p_control_estimate
#>   [1] 0.7441860 0.6865672 0.7317073 0.6526316 0.6842105 0.7087379 0.6875000
#>   [8] 0.7840909 0.7204301 0.6774194 0.8051948 0.7142857 0.6511628 0.7019231
#>  [15] 0.6237624 0.7500000 0.7802198 0.6565657 0.7894737 0.7040816 0.7641509
#>  [22] 0.6700000 0.7311828 0.6637168 0.8043478 0.6972477 0.6326531 0.7368421
#>  [29] 0.6976744 0.5913978 0.7070707 0.7033898 0.7184466 0.7543860 0.7500000
#>  [36] 0.7614679 0.7477477 0.7078652 0.6831683 0.6600000 0.7142857 0.6630435
#>  [43] 0.7090909 0.6947368 0.7307692 0.6969697 0.6160714 0.6597938 0.7000000
#>  [50] 0.7264151 0.7386364 0.6931818 0.7129630 0.6636364 0.7352941 0.6530612
#>  [57] 0.7446809 0.7113402 0.6836735 0.5871560 0.6168224 0.6448598 0.7216495
#>  [64] 0.6173913 0.7419355 0.6489362 0.6407767 0.6836735 0.6631579 0.6666667
#>  [71] 0.6036036 0.6500000 0.6590909 0.7669903 0.6931818 0.6739130 0.7613636
#>  [78] 0.6636364 0.7052632 0.6868687 0.5952381 0.7553191 0.6036036 0.6363636
#>  [85] 0.8163265 0.7525773 0.6746988 0.7735849 0.7777778 0.7096774 0.7241379
#>  [92] 0.7710843 0.6276596 0.7272727 0.6881720 0.7234043 0.6741573 0.6867470
#>  [99] 0.7291667 0.6129032
#> 
#> $p_treatment_estimate
#>   [1] 0.5614035 0.5864662 0.6186441 0.6571429 0.6627907 0.6804124 0.5666667
#>   [8] 0.5892857 0.6074766 0.6074766 0.5934959 0.6379310 0.6140351 0.6875000
#>  [15] 0.7474747 0.6166667 0.6697248 0.6831683 0.7333333 0.6666667 0.6702128
#>  [22] 0.6700000 0.7102804 0.7241379 0.5555556 0.6593407 0.7058824 0.6209677
#>  [29] 0.6052632 0.6355140 0.6435644 0.7073171 0.7113402 0.4651163 0.5000000
#>  [36] 0.7032967 0.6741573 0.6486486 0.6060606 0.6900000 0.6260163 0.7037037
#>  [43] 0.7000000 0.6285714 0.6562500 0.6435644 0.6704545 0.6019417 0.6666667
#>  [50] 0.7021277 0.6339286 0.6607143 0.5978261 0.6000000 0.7244898 0.7058824
#>  [57] 0.6509434 0.6893204 0.6176471 0.6043956 0.6881720 0.6989247 0.6699029
#>  [64] 0.7176471 0.6355140 0.6698113 0.7010309 0.6470588 0.6857143 0.6734694
#>  [71] 0.6629213 0.6900000 0.6696429 0.6701031 0.5446429 0.6759259 0.6607143
#>  [78] 0.6111111 0.6476190 0.6138614 0.5775862 0.6698113 0.6629213 0.6336634
#>  [85] 0.4509804 0.6699029 0.6068376 0.5319149 0.6847826 0.5981308 0.6991150
#>  [92] 0.6068376 0.6226415 0.6732673 0.6261682 0.7075472 0.6486486 0.5641026
#>  [99] 0.6923077 0.6074766
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 100 200 200 200 200 200 200 200 200 100 100 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 100 200 200 100 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1]  86  67  82  95 114 103  80  88  93  93  77  84  86 104 101  80  91  99
#>  [19]  95  98 106 100  93 113  46 109  98  76  86  93  99 118 103  57  52 109
#>  [37] 111  89 101 100  77  92 110  95 104  99 112  97 110 106  88  88 108 110
#>  [55] 102  98  94  97  98 109 107 107  97 115  93  94 103  98  95 102 111 100
#>  [73]  88 103  88  92  88 110  95  99  84  94 111  99  49  97  83  53 108  93
#>  [91]  87  83  94  99  93  94  89  83  96  93
#> 
#> $N_treatment
#>   [1] 114 133 118 105  86  97 120 112 107 107 123 116 114  96  99 120 109 101
#>  [19] 105 102  94 100 107  87  54  91 102 124 114 107 101  82  97  43  48  91
#>  [37]  89 111  99 100 123 108  90 105  96 101  88 103  90  94 112 112  92  90
#>  [55]  98 102 106 103 102  91  93  93 103  85 107 106  97 102 105  98  89 100
#>  [73] 112  97 112 108 112  90 105 101 116 106  89 101  51 103 117  47  92 107
#>  [91] 113 117 106 101 107 106 111 117 104 107
#> 

binomialfreq(p_control = 0.5, p_treatment = 0.40, N_total = 200,
            block_number = 2, simulation = 100, drift = -0.15)

#> $power
#> [1] 0.44
#> 
#> $p_control_estimate
#>   [1] 0.3962264 0.3291139 0.4235294 0.3789474 0.3663366 0.4505495 0.3917526
#>   [8] 0.4038462 0.3367347 0.4204545 0.4337349 0.3645833 0.4252874 0.3762376
#>  [15] 0.3789474 0.3655914 0.3750000 0.3814433 0.5333333 0.4117647 0.5102041
#>  [22] 0.4313725 0.4285714 0.3900000 0.4050633 0.5737705 0.3809524 0.4408602
#>  [29] 0.3888889 0.3510638 0.3908046 0.3473684 0.3620690 0.4000000 0.4166667
#>  [36] 0.3750000 0.3448276 0.3956044 0.3883495 0.3978495 0.4534884 0.3367347
#>  [43] 0.3465347 0.4875000 0.4146341 0.3979592 0.3168317 0.5102041 0.3673469
#>  [50] 0.3943662 0.2903226 0.4022989 0.4175824 0.4255319 0.4431818 0.4776119
#>  [57] 0.3953488 0.4556962 0.4215686 0.4059406 0.4273504 0.4056604 0.5918367
#>  [64] 0.3398058 0.4642857 0.3786408 0.3800000 0.3809524 0.3854167 0.4698795
#>  [71] 0.4615385 0.4148936 0.3863636 0.3277311 0.4090909 0.4196429 0.3592233
#>  [78] 0.3750000 0.4130435 0.3225806 0.3693694 0.4038462 0.3333333 0.4625000
#>  [85] 0.3495146 0.4494382 0.3367347 0.3773585 0.4938272 0.3736264 0.3617021
#>  [92] 0.2886598 0.4408602 0.3493976 0.5108696 0.4455446 0.4100000 0.4102564
#>  [99] 0.4040404 0.4220183
#> 
#> $p_treatment_estimate
#>   [1] 0.3829787 0.2809917 0.2782609 0.3333333 0.3131313 0.2568807 0.3398058
#>   [8] 0.3333333 0.3529412 0.3482143 0.2393162 0.3076923 0.2743363 0.3232323
#>  [15] 0.3142857 0.3738318 0.2980769 0.3495146 0.2727273 0.2173913 0.1960784
#>  [22] 0.2959184 0.3486239 0.2800000 0.3553719 0.2307692 0.2155172 0.3551402
#>  [29] 0.2187500 0.2735849 0.2743363 0.2952381 0.2976190 0.3700000 0.2403846
#>  [36] 0.2980769 0.3362832 0.2385321 0.2577320 0.2242991 0.2719298 0.3235294
#>  [43] 0.3232323 0.2833333 0.2881356 0.3333333 0.2424242 0.2352941 0.2549020
#>  [50] 0.2635659 0.3684211 0.3008850 0.3302752 0.3018868 0.2500000 0.2706767
#>  [57] 0.2192982 0.2809917 0.2959184 0.3232323 0.3855422 0.3829787 0.3333333
#>  [64] 0.3814433 0.3362069 0.3814433 0.2600000 0.2844828 0.2019231 0.3076923
#>  [71] 0.3020833 0.3490566 0.2767857 0.3703704 0.2857143 0.3522727 0.2989691
#>  [78] 0.3409091 0.2592593 0.2990654 0.4382022 0.3437500 0.2521008 0.2833333
#>  [85] 0.3195876 0.2792793 0.3333333 0.3085106 0.2521008 0.2477064 0.3018868
#>  [92] 0.2427184 0.2990654 0.2649573 0.3148148 0.3434343 0.2900000 0.2049180
#>  [99] 0.3564356 0.2967033
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 100 200 100 200 200 200 200 100 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 100 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 100 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1] 106  79  85  95 101  91  97 104  98  88  83  96  87 101  95  93  96  97
#>  [19]  45  85  49 102  91 100  79  61  84  93  72  94  87  95 116 100  96  96
#>  [37]  87  91 103  93  86  98 101  80  82  98 101  49  98  71 124  87  91  94
#>  [55]  88  67  86  79 102 101 117 106  49 103  84 103 100  84  96  83 104  94
#>  [73]  88 119  88 112 103 112  92  93 111 104  81  80 103  89  98 106  81  91
#>  [91]  94  97  93  83  92 101 100  78  99 109
#> 
#> $N_treatment
#>   [1]  94 121 115 105  99 109 103  96 102 112 117 104 113  99 105 107 104 103
#>  [19]  55 115  51  98 109 100 121  39 116 107 128 106 113 105  84 100 104 104
#>  [37] 113 109  97 107 114 102  99 120 118 102  99  51 102 129  76 113 109 106
#>  [55] 112 133 114 121  98  99  83  94  51  97 116  97 100 116 104 117  96 106
#>  [73] 112  81 112  88  97  88 108 107  89  96 119 120  97 111 102  94 119 109
#>  [91] 106 103 107 117 108  99 100 122 101  91
#>