Block Design for Response-Adaptive Randomization for Binomial Data

Simulation for binomial counts for block design for response-adaptive randomization with time as a confounding

binomialfreq(p_control, p_treatment, N_total, block_number = 4,
  drift = 0, simulation = 10000, zvalue = c(1, 1.5, 2),
  rand_ratio = c(1, 1.5, 2, 2.5), conf_int = 0.95,
  alternative = "less", correct = FALSE, replace = TRUE)

Arguments

p_control
scalar. Proportion of events under the control arm.

p_treatment
scalar. Proportion of events under the treatment arm.

N_total
scalar. Total sample size.

block_number
scalar. Number of blocks or time levels. The default is set to 4. If block_number is set to 1. This is a traditional RCT design.

drift
scalar. The increase or decrease in proportion of event over time. In this case, the proportion of failure changes in each block by the number of patient accured over the total sample size. The full drift effect is seen in the final block.

simulation
scalar. Number of simulation to be ran. The default is set to 10000.

zvalue
vector. The z-value cutoff for corrected chi-square test statistics.

rand_ratio
vector. The randomization ratio is set based on the corrected chi-square test statistic. The length of rand_ratio should be the same length of zvalue.

conf_int
scalar. Confidence level of the interval.

alternative
character. A string specifying the alternative hypothesis, must be one of "less" (default) or "greater".

correct
logical. a logical indicating whether to apply continuity correction when computing the test statistic: one half is subtracted from all |O - E| differences; however, the correction will not be bigger than the differences themselves.

replace
logical. should sampling be with replacement? If replace is set to FALSE (default), the 0 for control, 1 for treatment is replicated to the closest integer and this vector is sampled with no replacement. If replace is set to TRUE, the sampling is done based on randomization ratio provided with replacement.

Value

a list with details on the simulation.

power

scalar. The power of the trial, ie. the proportion of success over the number of simulation ran.

p_control_estimate

scalar. The estimated proportion of events under the control group.

p_treatment_estimate

scalar. The estimated proportion of events under the treatment group.

N_enrolled

vector. The number of patients enrolled in the trial (sum of control and experimental group for each simulation. )

N_control

vector. The number of patients enrolled in the control group for each simulation.

N_control

vector. The number of patients enrolled in the experimental group for each simulation.

Examples

binomialfreq(p_control = 0.7, p_treatment = 0.65, N_total = 200,
            block_number = 2, simulation = 100)

#> $power
#> [1] 0.24
#> 
#> $p_control_estimate
#>   [1] 0.6172840 0.7037037 0.6969697 0.7323944 0.7115385 0.7241379 0.7187500
#>   [8] 0.7250000 0.7027027 0.7272727 0.7162162 0.6813187 0.6464646 0.6989247
#>  [15] 0.6542056 0.6847826 0.7526882 0.7100000 0.7179487 0.6966292 0.7125000
#>  [22] 0.7835052 0.6960784 0.6857143 0.6464646 0.6581197 0.6250000 0.6593407
#>  [29] 0.7090909 0.6756757 0.7619048 0.7283951 0.6595745 0.7052632 0.7195122
#>  [36] 0.7282609 0.6666667 0.7100000 0.6666667 0.8155340 0.6526316 0.7397260
#>  [43] 0.7117117 0.6746988 0.6562500 0.7100000 0.7536232 0.6987952 0.6699029
#>  [50] 0.6292135 0.6938776 0.6060606 0.7000000 0.7070707 0.6703297 0.6476190
#>  [57] 0.7317073 0.6900000 0.6734694 0.6823529 0.7765957 0.6489362 0.6930693
#>  [64] 0.7619048 0.7500000 0.6276596 0.7802198 0.7500000 0.7524752 0.6703297
#>  [71] 0.7010309 0.8245614 0.6881720 0.7303371 0.6956522 0.7241379 0.7032967
#>  [78] 0.6363636 0.6315789 0.7126437 0.7777778 0.7052632 0.7096774 0.7450980
#>  [85] 0.7333333 0.7234043 0.7901235 0.7090909 0.7058824 0.6222222 0.7640449
#>  [92] 0.7117117 0.7586207 0.6782609 0.6458333 0.7456140 0.8200000 0.7640449
#>  [99] 0.6250000 0.7560976
#> 
#> $p_treatment_estimate
#>   [1] 0.5294118 0.7065217 0.6534653 0.6201550 0.7083333 0.6548673 0.6730769
#>   [8] 0.6500000 0.6966292 0.6071429 0.5317460 0.5596330 0.7227723 0.6542056
#>  [15] 0.6451613 0.6851852 0.7102804 0.7100000 0.6147541 0.5405405 0.5500000
#>  [22] 0.7087379 0.6530612 0.6842105 0.7425743 0.6987952 0.6979167 0.6513761
#>  [29] 0.6777778 0.6404494 0.6034483 0.5882353 0.6792453 0.6285714 0.6271186
#>  [36] 0.6666667 0.6542056 0.6600000 0.5769231 0.6804124 0.6571429 0.6377953
#>  [43] 0.6179775 0.6666667 0.6442308 0.7200000 0.6641221 0.6239316 0.6804124
#>  [50] 0.5765766 0.6764706 0.6237624 0.6636364 0.7029703 0.7064220 0.6210526
#>  [57] 0.5762712 0.6000000 0.6862745 0.6695652 0.6226415 0.6886792 0.7272727
#>  [64] 0.6810345 0.6847826 0.6886792 0.6055046 0.7159091 0.6868687 0.6605505
#>  [71] 0.6796117 0.5813953 0.6261682 0.5855856 0.7058824 0.6017699 0.7522936
#>  [78] 0.6732673 0.6476190 0.5840708 0.5090909 0.5714286 0.7102804 0.6326531
#>  [85] 0.7052632 0.6509434 0.6638655 0.6222222 0.6020408 0.5454545 0.6576577
#>  [92] 0.6966292 0.6371681 0.7058824 0.6826923 0.7209302 0.5400000 0.6576577
#>  [99] 0.6250000 0.5932203
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 100
#>  [73] 200 200 200 200 200 200 200 200 100 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 100 200 200 200
#> 
#> $N_control
#>   [1]  81 108  99  71 104  87  96 120 111  88  74  91  99  93 107  92  93 100
#>  [19]  78  89  80  97 102 105  99 117 104  91 110 111  84  81  94  95  82  92
#>  [37]  93 100  96 103  95  73 111  83  96 100  69  83 103  89  98  99  90  99
#>  [55]  91 105  82 100  98  85  94  94 101  84 108  94  91 112 101  91  97  57
#>  [73]  93  89 115  87  91  99  95  87  45  95  93 102 105  94  81 110 102  90
#>  [91]  89 111  87 115  96 114  50  89  88  82
#> 
#> $N_treatment
#>   [1] 119  92 101 129  96 113 104  80  89 112 126 109 101 107  93 108 107 100
#>  [19] 122 111 120 103  98  95 101  83  96 109  90  89 116 119 106 105 118 108
#>  [37] 107 100 104  97 105 127  89 117 104 100 131 117  97 111 102 101 110 101
#>  [55] 109  95 118 100 102 115 106 106  99 116  92 106 109  88  99 109 103  43
#>  [73] 107 111  85 113 109 101 105 113  55 105 107  98  95 106 119  90  98 110
#>  [91] 111  89 113  85 104  86  50 111 112 118
#> 

binomialfreq(p_control = 0.5, p_treatment = 0.40, N_total = 200,
            block_number = 2, simulation = 100, drift = -0.15)

#> $power
#> [1] 0.44
#> 
#> $p_control_estimate
#>   [1] 0.3855422 0.4130435 0.3750000 0.4864865 0.3690476 0.3414634 0.3804348
#>   [8] 0.3763441 0.3877551 0.3097345 0.4271845 0.4456522 0.4065934 0.3333333
#>  [15] 0.4722222 0.4050633 0.5540541 0.3398058 0.4268293 0.4567901 0.4516129
#>  [22] 0.3513514 0.4318182 0.3333333 0.3714286 0.3364486 0.4056604 0.3000000
#>  [29] 0.3763441 0.3428571 0.4148936 0.4000000 0.4235294 0.3918919 0.4040404
#>  [36] 0.4933333 0.5625000 0.4226804 0.3932584 0.3809524 0.3734940 0.4050633
#>  [43] 0.4040404 0.4430380 0.3761468 0.3723404 0.3800000 0.4731183 0.5102041
#>  [50] 0.4363636 0.4745763 0.4021739 0.3534483 0.4631579 0.3370787 0.4411765
#>  [57] 0.5357143 0.3695652 0.2916667 0.4095238 0.5555556 0.4197531 0.4777778
#>  [64] 0.3465347 0.4285714 0.3157895 0.3875000 0.3979592 0.4024390 0.4935065
#>  [71] 0.4117647 0.4404762 0.4578313 0.3076923 0.3789474 0.4368932 0.4594595
#>  [78] 0.3103448 0.4196429 0.4300000 0.4020619 0.3232323 0.4174757 0.4000000
#>  [85] 0.3793103 0.3797468 0.4705882 0.3174603 0.2954545 0.4418605 0.3800000
#>  [92] 0.3684211 0.3092784 0.3561644 0.3058824 0.3469388 0.4183673 0.3571429
#>  [99] 0.3047619 0.3818182
#> 
#> $p_treatment_estimate
#>   [1] 0.2307692 0.3148148 0.2265625 0.2698413 0.2327586 0.2711864 0.3148148
#>   [8] 0.2990654 0.3137255 0.3333333 0.3505155 0.2314815 0.2844037 0.1896552
#>  [15] 0.3046875 0.2644628 0.2777778 0.2886598 0.2711864 0.2857143 0.2616822
#>  [22] 0.2380952 0.2232143 0.2758621 0.3052632 0.3333333 0.2872340 0.2600000
#>  [29] 0.2803738 0.3789474 0.3207547 0.3777778 0.2782609 0.2777778 0.3267327
#>  [36] 0.2480000 0.2500000 0.3398058 0.2792793 0.2189781 0.2222222 0.2975207
#>  [43] 0.3564356 0.2148760 0.2967033 0.2547170 0.3000000 0.2897196 0.2549020
#>  [50] 0.3666667 0.1951220 0.2870370 0.2976190 0.2952381 0.2432432 0.3181818
#>  [57] 0.2500000 0.2592593 0.2788462 0.4000000 0.2968750 0.2268908 0.2545455
#>  [64] 0.3232323 0.1568627 0.2857143 0.2750000 0.2941176 0.2796610 0.3008130
#>  [71] 0.3163265 0.2931034 0.2564103 0.3119266 0.3809524 0.3298969 0.3483146
#>  [78] 0.2920354 0.3977273 0.3200000 0.2718447 0.2475248 0.2783505 0.3052632
#>  [85] 0.2743363 0.2892562 0.1836735 0.2335766 0.2946429 0.3245614 0.3100000
#>  [92] 0.3428571 0.2718447 0.2047244 0.2347826 0.3137255 0.3137255 0.2241379
#>  [99] 0.2736842 0.3555556
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 100 200 200 200 200 200 200 200 200 200 200 200 100 200 100 200 200 200
#>  [55] 200 200 100 200 200 200 200 200 200 200 100 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 100 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1]  83  92  72  74  84  82  92  93  98 113 103  92  91  84  72  79  74 103
#>  [19]  82  81  93  74  88  84 105 107 106 100  93 105  94 110  85  74  99  75
#>  [37]  48  97  89  63  83  79  99  79 109  94 100  93  49 110  59  92 116  95
#>  [55]  89  68  56  92  96 105  72  81  90 101  49  95  80  98  82  77 102  84
#>  [73]  83  91  95 103 111  87 112 100  97  99 103 105  87  79  51  63  88  86
#>  [91] 100  95  97  73  85  98  98  84 105 110
#> 
#> $N_treatment
#>   [1] 117 108 128 126 116 118 108 107 102  87  97 108 109 116 128 121 126  97
#>  [19] 118 119 107 126 112 116  95  93  94 100 107  95 106  90 115 126 101 125
#>  [37]  52 103 111 137 117 121 101 121  91 106 100 107  51  90  41 108  84 105
#>  [55] 111 132  44 108 104  95 128 119 110  99  51 105 120 102 118 123  98 116
#>  [73] 117 109 105  97  89 113  88 100 103 101  97  95 113 121  49 137 112 114
#>  [91] 100 105 103 127 115 102 102 116  95  90
#>