Block Design for Response-Adaptive Randomization for Binomial Data

Simulation for binomial counts for block design for response-adaptive randomization with time as a confounding

binomialfreq(p_control, p_treatment, N_total, block_number = 4,
  drift = 0, simulation = 10000, conf_int = 0.95,
  alternative = "greater", correct = FALSE, replace = TRUE,
  early_stop = FALSE, size_equal_randomization = 20,
  min_patient_earlystop = 20, max_prob = 0.8)

Arguments

p_control
scalar. Proportion of events under the control arm.

p_treatment
scalar. Proportion of events under the treatment arm.

N_total
scalar. Total sample size.

block_number
scalar. Number of blocks or time levels. The default is set to 4. If block_number is set to 1. This is a traditional RCT design.

drift
scalar. The increase or decrease in proportion of event over time. In this case, the proportion of failure changes in each block by the number of patient accured over the total sample size. The full drift effect is seen in the final block.

simulation
scalar. Number of simulation to be ran. The default is set to 10000.

conf_int
scalar. Confidence level of the interval.

alternative
character. A string specifying the alternative hypothesis, must be one of "less" or "greater" (default).

correct
logical. A logical indicating whether to apply continuity correction when computing the test statistic: one half is subtracted from all |O - E| differences; however, the correction will not be bigger than the differences themselves.

replace
logical. should sampling be with replacement? If replace is set to FALSE (default), the 0 for control, 1 for treatment is replicated to the closest integer and this vector is sampled with no replacement. If replace is set to TRUE, the sampling is done based on randomization ratio provided with replacement.

early_stop
logical. A logical indicating whether the trials are stopped early for success or futility.

size_equal_randomization
scalar. The number of run in patients because adaptive randomization is applied.

min_patient_earlystop
scalar. Minimum number of patients before early stopping rule is applied.

max_prob
scalar. The maximum probability for assigning to treatment/control group is 0.8.

Value

a list with details on the simulation.

power

scalar. The power of the trial, ie. the proportion of success over the number of simulation ran.

p_control_estimate

scalar. The estimated proportion of events under the control group.

p_treatment_estimate

scalar. The estimated proportion of events under the treatment group.

N_enrolled

vector. The number of patients enrolled in the trial (sum of control and experimental group for each simulation. )

N_control

vector. The number of patients enrolled in the control group for each simulation.

N_control

vector. The number of patients enrolled in the experimental group for each simulation.

randomization_ratio

matrix. The randomization ratio allocated for each block.

Examples

binomialfreq(p_control = 0.7, p_treatment = 0.65, N_total = 200,
            block_number = 2, simulation = 100)

#> $power
#> [1] 0.01
#> 
#> $p_control_estimate
#>   [1] 0.6914894 0.7280702 0.6800000 0.7727273 0.6504854 0.7009346 0.6754386
#>   [8] 0.5979381 0.6923077 0.7157895 0.6703297 0.7428571 0.7217391 0.6138614
#>  [15] 0.7043478 0.6607143 0.6979167 0.7525773 0.6516854 0.7200000 0.7415730
#>  [22] 0.6391753 0.7454545 0.6458333 0.7641509 0.7610619 0.6407767 0.6605505
#>  [29] 0.6764706 0.7000000 0.7500000 0.7142857 0.6947368 0.7500000 0.7731959
#>  [36] 0.6565657 0.6600000 0.6698113 0.7000000 0.7169811 0.6517857 0.7142857
#>  [43] 0.6875000 0.7551020 0.7745098 0.6736842 0.6595745 0.7211538 0.7187500
#>  [50] 0.7358491 0.6000000 0.7674419 0.6842105 0.6696429 0.7090909 0.7340426
#>  [57] 0.8090909 0.6728972 0.6792453 0.7029703 0.7913043 0.7400000 0.7500000
#>  [64] 0.6785714 0.7545455 0.6086957 0.7567568 0.7113402 0.6435644 0.7191011
#>  [71] 0.7115385 0.6759259 0.6627907 0.7065217 0.7008547 0.7403846 0.6435644
#>  [78] 0.6785714 0.7864078 0.6700000 0.7200000 0.7431193 0.6630435 0.7256637
#>  [85] 0.7524752 0.7113402 0.6700000 0.7777778 0.6236559 0.6593407 0.6966292
#>  [92] 0.7663551 0.6504854 0.7317073 0.6111111 0.6962025 0.6846847 0.6767677
#>  [99] 0.6851852 0.7394958
#> 
#> $p_treatment_estimate
#>   [1] 0.6509434 0.6279070 0.6300000 0.6333333 0.5876289 0.6129032 0.6395349
#>   [8] 0.7087379 0.6146789 0.6000000 0.6880734 0.6526316 0.6470588 0.6464646
#>  [15] 0.6235294 0.6136364 0.6442308 0.6407767 0.7117117 0.6700000 0.6666667
#>  [22] 0.6310680 0.6444444 0.5961538 0.6595745 0.6896552 0.7010309 0.6153846
#>  [29] 0.6836735 0.6200000 0.6590909 0.7254902 0.6952381 0.5909091 0.6601942
#>  [36] 0.7029703 0.6300000 0.6382979 0.6300000 0.6914894 0.6590909 0.6631579
#>  [43] 0.6538462 0.6666667 0.6224490 0.6666667 0.6037736 0.6770833 0.6250000
#>  [50] 0.7127660 0.6181818 0.6052632 0.6666667 0.5795455 0.6333333 0.6320755
#>  [57] 0.6777778 0.6666667 0.6170213 0.7171717 0.6000000 0.6900000 0.6666667
#>  [64] 0.6022727 0.5444444 0.6296296 0.6179775 0.6504854 0.6262626 0.6936937
#>  [71] 0.6354167 0.5760870 0.7280702 0.6388889 0.6987952 0.6979167 0.6161616
#>  [78] 0.7272727 0.7113402 0.6700000 0.6200000 0.6923077 0.6666667 0.6206897
#>  [85] 0.6161616 0.5631068 0.6400000 0.6237624 0.6915888 0.6788991 0.6936937
#>  [92] 0.6451613 0.6597938 0.5974026 0.6818182 0.6611570 0.6629213 0.6237624
#>  [99] 0.5978261 0.6049383
#> 
#> $prop_diff_estimate
#>   [1] -0.0373944511 -0.1036636137 -0.0500000000 -0.1388101983 -0.0792299899
#>   [6] -0.0869300912 -0.0359624029  0.1130034061 -0.0768142308 -0.0999380549
#>  [11]  0.0185521274 -0.0891029500 -0.0801740211  0.0338338338 -0.0827360229
#>  [16] -0.0466910272 -0.0526739043 -0.1115793699  0.0610376399 -0.0575757576
#>  [21] -0.0751164203 -0.0148645897 -0.0952960260 -0.0582760016 -0.1044444444
#>  [26] -0.0756923077  0.0603085554 -0.0569503980  0.0072492952 -0.0821256039
#>  [31] -0.0910569106  0.0112044818 -0.0020100503 -0.1599675193 -0.1116008816
#>  [36]  0.0482482482 -0.0280448718 -0.0287915653 -0.0720594139 -0.0248995984
#>  [41]  0.0052845528 -0.0535821794 -0.0336673347 -0.0884353741 -0.1537229784
#>  [46] -0.0110552764 -0.0586345382 -0.0457463884 -0.1229813665 -0.0230830992
#>  [51]  0.0182112505 -0.1566364386 -0.0058621387 -0.0910209102 -0.0734693878
#>  [56] -0.1029352634 -0.1285714286 -0.0058291457 -0.0622240064  0.0070578746
#>  [61] -0.1863874346 -0.0502200880 -0.0806060606 -0.0801558016 -0.2112505059
#>  [66]  0.0328542094 -0.1401831129 -0.0724607330 -0.0158158158 -0.0259161774
#>  [71] -0.0763527054 -0.0978654853  0.0629342051 -0.0534846029 -0.0051535766
#>  [76] -0.0424679487 -0.0219095477  0.0491270808 -0.0782661048 -0.0032051282
#>  [81] -0.0972388956 -0.0484359233  0.0049019608 -0.1043098064 -0.1401644275
#>  [86] -0.1412714430 -0.0348837209 -0.1619532045  0.0681407035  0.0210739615
#>  [91]  0.0004099201 -0.1276595745  0.0098098098 -0.1266310160  0.0697485807
#>  [96] -0.0325152087 -0.0297049847 -0.0516233112 -0.0882828283 -0.1312669024
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1]  94 114 100 110 103 107 114  97  91  95  91 105 115 101 115 112  96  97
#>  [19]  89 100  89  97 110  96 106 113 103 109 102 100 112  98  95 112  97  99
#>  [37] 100 106 100 106 112 105  96  98 102  95  94 104  96 106  90  86  95 112
#>  [55] 110  94 110 107 106 101 115 100  92 112 110  92 111  97 101  89 104 108
#>  [73]  86  92 117 104 101 112 103 100 100 109  92 113 101  97 100  99  93  91
#>  [91]  89 107 103 123  90  79 111  99 108 119
#> 
#> $N_treatment
#>   [1] 106  86 100  90  97  93  86 103 109 105 109  95  85  99  85  88 104 103
#>  [19] 111 100 111 103  90 104  94  87  97  91  98 100  88 102 105  88 103 101
#>  [37] 100  94 100  94  88  95 104 102  98 105 106  96 104  94 110 114 105  88
#>  [55]  90 106  90  93  94  99  85 100 108  88  90 108  89 103  99 111  96  92
#>  [73] 114 108  83  96  99  88  97 100 100  91 108  87  99 103 100 101 107 109
#>  [91] 111  93  97  77 110 121  89 101  92  81
#> 
#> $early_stop
#>   [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [38] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [75] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $randomization_ratio
#>        [,1]      [,2]
#>   [1,]  0.5 0.4748886
#>   [2,]  0.5 0.4225352
#>   [3,]  0.5 0.5147059
#>   [4,]  0.5 0.4449838
#>   [5,]  0.5 0.4826087
#>   [6,]  0.5 0.4531384
#>   [7,]  0.5 0.5039932
#>   [8,]  0.5 0.5181269
#>   [9,]  0.5 0.5102041
#>  [10,]  0.5 0.4414186
#>  [11,]  0.5 0.5591910
#>  [12,]  0.5 0.4523678
#>  [13,]  0.5 0.4573849
#>  [14,]  0.5 0.5142071
#>  [15,]  0.5 0.4638078
#>  [16,]  0.5 0.4628571
#>  [17,]  0.5 0.4960068
#>  [18,]  0.5 0.4685466
#>  [19,]  0.5 0.5585831
#>  [20,]  0.5 0.4585366
#>  [21,]  0.5 0.4763675
#>  [22,]  0.5 0.4520548
#>  [23,]  0.5 0.4518177
#>  [24,]  0.5 0.5214368
#>  [25,]  0.5 0.4425454
#>  [26,]  0.5 0.4736842
#>  [27,]  0.5 0.5216179
#>  [28,]  0.5 0.4692828
#>  [29,]  0.5 0.4908333
#>  [30,]  0.5 0.4832215
#>  [31,]  0.5 0.4626866
#>  [32,]  0.5 0.4698605
#>  [33,]  0.5 0.4924044
#>  [34,]  0.5 0.4499183
#>  [35,]  0.5 0.4501486
#>  [36,]  0.5 0.5031627
#>  [37,]  0.5 0.4963899
#>  [38,]  0.5 0.4807692
#>  [39,]  0.5 0.4427711
#>  [40,]  0.5 0.4842767
#>  [41,]  0.5 0.4375000
#>  [42,]  0.5 0.4558140
#>  [43,]  0.5 0.5074167
#>  [44,]  0.5 0.4775752
#>  [45,]  0.5 0.4605263
#>  [46,]  0.5 0.4923077
#>  [47,]  0.5 0.4980080
#>  [48,]  0.5 0.4955752
#>  [49,]  0.5 0.4722513
#>  [50,]  0.5 0.5071249
#>  [51,]  0.5 0.4949495
#>  [52,]  0.5 0.4602841
#>  [53,]  0.5 0.4720765
#>  [54,]  0.5 0.4402255
#>  [55,]  0.5 0.4603084
#>  [56,]  0.5 0.4589097
#>  [57,]  0.5 0.4266667
#>  [58,]  0.5 0.5064748
#>  [59,]  0.5 0.4794334
#>  [60,]  0.5 0.4678743
#>  [61,]  0.5 0.4142857
#>  [62,]  0.5 0.5301395
#>  [63,]  0.5 0.5103926
#>  [64,]  0.5 0.4898911
#>  [65,]  0.5 0.4339250
#>  [66,]  0.5 0.4994325
#>  [67,]  0.5 0.4200861
#>  [68,]  0.5 0.4525386
#>  [69,]  0.5 0.4824535
#>  [70,]  0.5 0.5091667
#>  [71,]  0.5 0.4380165
#>  [72,]  0.5 0.4763231
#>  [73,]  0.5 0.5769341
#>  [74,]  0.5 0.4559585
#>  [75,]  0.5 0.4957983
#>  [76,]  0.5 0.4807122
#>  [77,]  0.5 0.4729730
#>  [78,]  0.5 0.4917989
#>  [79,]  0.5 0.4842767
#>  [80,]  0.5 0.5269072
#>  [81,]  0.5 0.5227618
#>  [82,]  0.5 0.4715155
#>  [83,]  0.5 0.5264624
#>  [84,]  0.5 0.4253545
#>  [85,]  0.5 0.4613827
#>  [86,]  0.5 0.4667089
#>  [87,]  0.5 0.4831873
#>  [88,]  0.5 0.4460094
#>  [89,]  0.5 0.5441860
#>  [90,]  0.5 0.4903846
#>  [91,]  0.5 0.4931507
#>  [92,]  0.5 0.4247855
#>  [93,]  0.5 0.4865265
#>  [94,]  0.5 0.4346523
#>  [95,]  0.5 0.5734597
#>  [96,]  0.5 0.5065789
#>  [97,]  0.5 0.4724653
#>  [98,]  0.5 0.4645309
#>  [99,]  0.5 0.4572908
#> [100,]  0.5 0.4383358
#> 

binomialfreq(p_control = 0.5, p_treatment = 0.40, N_total = 200,
            block_number = 2, simulation = 100, drift = -0.15)

#> $power
#> [1] 0
#> 
#> $p_control_estimate
#>   [1] 0.3750000 0.4687500 0.4553571 0.3883495 0.3250000 0.4666667 0.3478261
#>   [8] 0.4695652 0.4296875 0.4056604 0.4311927 0.3783784 0.5052632 0.3564356
#>  [15] 0.4500000 0.4591837 0.3592233 0.4000000 0.3584906 0.4476190 0.4270833
#>  [22] 0.4150943 0.4464286 0.4750000 0.4414414 0.4036697 0.3762376 0.2903226
#>  [29] 0.3839286 0.4554455 0.4485981 0.3921569 0.4300000 0.5158730 0.3783784
#>  [36] 0.4554455 0.4854369 0.3846154 0.5217391 0.4954128 0.4000000 0.4891304
#>  [43] 0.5054945 0.3921569 0.4607843 0.5038168 0.4695652 0.4513274 0.4196429
#>  [50] 0.4144144 0.4421053 0.4128440 0.4090909 0.4456522 0.3771930 0.4183673
#>  [57] 0.4700000 0.4695652 0.3908046 0.4395604 0.3333333 0.4271845 0.3500000
#>  [64] 0.3592233 0.4343434 0.4579439 0.4285714 0.4159292 0.4285714 0.3888889
#>  [71] 0.4444444 0.4273504 0.4166667 0.3761468 0.2903226 0.4166667 0.4245283
#>  [78] 0.3738318 0.5000000 0.3960396 0.4453782 0.3461538 0.3333333 0.3883495
#>  [85] 0.3853211 0.3949580 0.4250000 0.4607843 0.4727273 0.4485981 0.4473684
#>  [92] 0.4141414 0.4038462 0.5000000 0.4495413 0.5148515 0.3225806 0.5140187
#>  [99] 0.4300000 0.4268293
#> 
#> $p_treatment_estimate
#>   [1] 0.4107143 0.3269231 0.2840909 0.3402062 0.3250000 0.3000000 0.3055556
#>   [8] 0.3764706 0.1944444 0.3510638 0.3076923 0.3370787 0.3619048 0.2222222
#>  [15] 0.3600000 0.3725490 0.2680412 0.3809524 0.2872340 0.3263158 0.3653846
#>  [22] 0.3510638 0.2954545 0.3875000 0.2696629 0.3626374 0.2525253 0.3177570
#>  [29] 0.3181818 0.3131313 0.3118280 0.2959184 0.3200000 0.2027027 0.2471910
#>  [36] 0.3333333 0.3917526 0.3437500 0.2000000 0.2747253 0.4210526 0.4537037
#>  [43] 0.4128440 0.2755102 0.3469388 0.2608696 0.2470588 0.3333333 0.2840909
#>  [50] 0.3595506 0.4000000 0.3406593 0.3555556 0.4074074 0.3255814 0.3529412
#>  [57] 0.4000000 0.2705882 0.3716814 0.3486239 0.3097345 0.2886598 0.2900000
#>  [64] 0.3608247 0.4158416 0.2903226 0.3235294 0.3218391 0.2842105 0.3454545
#>  [71] 0.4782609 0.3012048 0.2625000 0.3736264 0.3738318 0.3478261 0.3723404
#>  [78] 0.3118280 0.3518519 0.3333333 0.3580247 0.4166667 0.2616822 0.2989691
#>  [85] 0.3186813 0.3456790 0.3000000 0.3163265 0.2888889 0.3225806 0.3139535
#>  [92] 0.3168317 0.3645833 0.3297872 0.2857143 0.3838384 0.3644860 0.3763441
#>  [99] 0.3900000 0.3898305
#> 
#> $prop_diff_estimate
#>   [1]  0.0345528455 -0.1453558504 -0.1892339545 -0.0275582573  0.0006273526
#>   [6] -0.1728187919 -0.0411290323 -0.0941272764 -0.2476963581 -0.0733982157
#>  [11] -0.1248235531 -0.0443540183 -0.1413675556 -0.1335472228 -0.0910364146
#>  [16] -0.1037352822 -0.0848886213 -0.0140703518 -0.0722891566 -0.1095613503
#>  [21] -0.0615230461 -0.0636658556 -0.1483333333 -0.1130434783 -0.1751824818
#>  [26] -0.0405268490 -0.1277277277  0.0353463947 -0.0678207739 -0.1431722478
#>  [31] -0.1538145178 -0.1008032129 -0.1227858293 -0.3244444444 -0.1418323427
#>  [36] -0.1165165165 -0.0944944945 -0.0408653846 -0.3280220948 -0.2490200124
#>  [41]  0.0293697736 -0.0447638604 -0.0982564103 -0.1107518189 -0.1152304609
#>  [46] -0.2364277321 -0.2252939963 -0.1189743590 -0.1440993789 -0.0580474934
#>  [51] -0.0369921164 -0.0336290489 -0.0734693878 -0.0427419355 -0.0708418891
#>  [56] -0.0698795181 -0.0791666667 -0.2254840725 -0.0212115032 -0.0418594404
#>  [61] -0.0147824276 -0.1422560609 -0.0681444992  0.0296846011 -0.0304496874
#>  [66] -0.1781155015 -0.1199756395 -0.1078451666 -0.1556886228 -0.0515789474
#>  [71]  0.0306451613 -0.1296639868 -0.1552453469  0.0090964221  0.0942249240
#>  [76] -0.0688405797 -0.0832303255 -0.0653791994 -0.1507293355 -0.0672672673
#>  [81] -0.0893042575  0.0666131621 -0.0725628141 -0.0882882883 -0.0712411705
#>  [86] -0.0384855419 -0.1384228188 -0.1473469388 -0.1930251419 -0.1288753799
#>  [91] -0.1273938385 -0.0900340886 -0.0472456775 -0.1820762368 -0.1249208694
#>  [96] -0.1233206336  0.0485593391 -0.1349989925 -0.0497792051 -0.0551867220
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1]  88  96 112 103  80 120  92 115 128 106 109 111  95 101 100  98 103  95
#>  [19] 106 105  96 106 112 120 111 109 101  93 112 101 107 102 100 126 111 101
#>  [37] 103 104 115 109 105  92  91 102 102 131 115 113 112 111  95 109 110  92
#>  [55] 114  98 100 115  87  91  87 103 100 103  99 107  98 113 105  90 108 117
#>  [73] 120 109  93 108 106 107  92 101 119 104  93 103 109 119 120 102 110 107
#>  [91] 114  99 104 106 109 101  93 107 100  82
#> 
#> $N_treatment
#>   [1] 112 104  88  97 120  80 108  85  72  94  91  89 105  99 100 102  97 105
#>  [19]  94  95 104  94  88  80  89  91  99 107  88  99  93  98 100  74  89  99
#>  [37]  97  96  85  91  95 108 109  98  98  69  85  87  88  89 105  91  90 108
#>  [55]  86 102 100  85 113 109 113  97 100  97 101  93 102  87  95 110  92  83
#>  [73]  80  91 107  92  94  93 108  99  81  96 107  97  91  81  80  98  90  93
#>  [91]  86 101  96  94  91  99 107  93 100 118
#> 
#> $early_stop
#>   [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [38] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [75] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $randomization_ratio
#>        [,1]      [,2]
#>   [1,]  0.5 0.5625000
#>   [2,]  0.5 0.3634021
#>   [3,]  0.5 0.3908046
#>   [4,]  0.5 0.5048900
#>   [5,]  0.5 0.5938038
#>   [6,]  0.5 0.3387242
#>   [7,]  0.5 0.5183867
#>   [8,]  0.5 0.4031350
#>   [9,]  0.5 0.3698838
#>  [10,]  0.5 0.4098361
#>  [11,]  0.5 0.4508475
#>  [12,]  0.5 0.3932039
#>  [13,]  0.5 0.4716981
#>  [14,]  0.5 0.4661017
#>  [15,]  0.5 0.4792824
#>  [16,]  0.5 0.4524362
#>  [17,]  0.5 0.5384615
#>  [18,]  0.5 0.5000000
#>  [19,]  0.5 0.4786236
#>  [20,]  0.5 0.4836066
#>  [21,]  0.5 0.5323571
#>  [22,]  0.5 0.4558541
#>  [23,]  0.5 0.4284274
#>  [24,]  0.5 0.4509804
#>  [25,]  0.5 0.3675000
#>  [26,]  0.5 0.4595376
#>  [27,]  0.5 0.3908046
#>  [28,]  0.5 0.6052632
#>  [29,]  0.5 0.4644809
#>  [30,]  0.5 0.3960857
#>  [31,]  0.5 0.4285714
#>  [32,]  0.5 0.4310345
#>  [33,]  0.5 0.4509804
#>  [34,]  0.5 0.2927757
#>  [35,]  0.5 0.4473684
#>  [36,]  0.5 0.4700984
#>  [37,]  0.5 0.4695686
#>  [38,]  0.5 0.4762955
#>  [39,]  0.5 0.3164557
#>  [40,]  0.5 0.3409354
#>  [41,]  0.5 0.4769765
#>  [42,]  0.5 0.4507202
#>  [43,]  0.5 0.5067985
#>  [44,]  0.5 0.4306220
#>  [45,]  0.5 0.4224063
#>  [46,]  0.5 0.3333333
#>  [47,]  0.5 0.3508342
#>  [48,]  0.5 0.3992606
#>  [49,]  0.5 0.3908046
#>  [50,]  0.5 0.3673197
#>  [51,]  0.5 0.5298446
#>  [52,]  0.5 0.5121951
#>  [53,]  0.5 0.4400000
#>  [54,]  0.5 0.5023401
#>  [55,]  0.5 0.4570637
#>  [56,]  0.5 0.4842105
#>  [57,]  0.5 0.4141331
#>  [58,]  0.5 0.3763314
#>  [59,]  0.5 0.4845070
#>  [60,]  0.5 0.5767442
#>  [61,]  0.5 0.5500778
#>  [62,]  0.5 0.4090909
#>  [63,]  0.5 0.3697479
#>  [64,]  0.5 0.4869110
#>  [65,]  0.5 0.4616547
#>  [66,]  0.5 0.4349680
#>  [67,]  0.5 0.4436348
#>  [68,]  0.5 0.3863636
#>  [69,]  0.5 0.3695652
#>  [70,]  0.5 0.4792746
#>  [71,]  0.5 0.5000000
#>  [72,]  0.5 0.4197421
#>  [73,]  0.5 0.3304721
#>  [74,]  0.5 0.4986554
#>  [75,]  0.5 0.5704258
#>  [76,]  0.5 0.4729730
#>  [77,]  0.5 0.4229471
#>  [78,]  0.5 0.4475138
#>  [79,]  0.5 0.4651163
#>  [80,]  0.5 0.4686717
#>  [81,]  0.5 0.4210526
#>  [82,]  0.5 0.5555556
#>  [83,]  0.5 0.4615385
#>  [84,]  0.5 0.4715074
#>  [85,]  0.5 0.4506253
#>  [86,]  0.5 0.4295302
#>  [87,]  0.5 0.4390244
#>  [88,]  0.5 0.4303087
#>  [89,]  0.5 0.4774920
#>  [90,]  0.5 0.3483607
#>  [91,]  0.5 0.4210526
#>  [92,]  0.5 0.4974874
#>  [93,]  0.5 0.4373214
#>  [94,]  0.5 0.3897116
#>  [95,]  0.5 0.4826271
#>  [96,]  0.5 0.4649524
#>  [97,]  0.5 0.5720054
#>  [98,]  0.5 0.4566929
#>  [99,]  0.5 0.5005675
#> [100,]  0.5 0.4985755
#>