Block Design for Response-Adaptive Randomization for Binomial Data

Simulation for binomial counts for block design for response-adaptive randomization with time as a confounding

binomialfreq(p_control, p_treatment, N_total, block_number = 4,
  drift = 0, simulation = 10000, conf_int = 0.95,
  alternative = "greater", correct = FALSE, replace = TRUE,
  early_stop = FALSE)

Arguments

p_control
scalar. Proportion of events under the control arm.

p_treatment
scalar. Proportion of events under the treatment arm.

N_total
scalar. Total sample size.

block_number
scalar. Number of blocks or time levels. The default is set to 4. If block_number is set to 1. This is a traditional RCT design.

drift
scalar. The increase or decrease in proportion of event over time. In this case, the proportion of failure changes in each block by the number of patient accured over the total sample size. The full drift effect is seen in the final block.

simulation
scalar. Number of simulation to be ran. The default is set to 10000.

conf_int
scalar. Confidence level of the interval.

alternative
character. A string specifying the alternative hypothesis, must be one of "less" or "greater" (default).

correct
logical. A logical indicating whether to apply continuity correction when computing the test statistic: one half is subtracted from all |O - E| differences; however, the correction will not be bigger than the differences themselves.

replace
logical. should sampling be with replacement? If replace is set to FALSE (default), the 0 for control, 1 for treatment is replicated to the closest integer and this vector is sampled with no replacement. If replace is set to TRUE, the sampling is done based on randomization ratio provided with replacement.

early_stop
logical. A logical indicating whether the trials are stopped early for success or futility.

Value

a list with details on the simulation.

power

scalar. The power of the trial, ie. the proportion of success over the number of simulation ran.

p_control_estimate

scalar. The estimated proportion of events under the control group.

p_treatment_estimate

scalar. The estimated proportion of events under the treatment group.

N_enrolled

vector. The number of patients enrolled in the trial (sum of control and experimental group for each simulation. )

N_control

vector. The number of patients enrolled in the control group for each simulation.

N_control

vector. The number of patients enrolled in the experimental group for each simulation.

Examples

binomialfreq(p_control = 0.7, p_treatment = 0.65, N_total = 200,
            block_number = 2, simulation = 100)

#> $power
#> [1] 0.01
#> 
#> $p_control_estimate
#>   [1] 0.7684211 0.6442308 0.7291667 0.6530612 0.7272727 0.6363636 0.6923077
#>   [8] 0.7450980 0.7500000 0.6744186 0.6388889 0.8265306 0.6603774 0.7578947
#>  [15] 0.6900000 0.7272727 0.6548673 0.7115385 0.7204301 0.7058824 0.6477273
#>  [22] 0.6545455 0.6808511 0.6458333 0.7755102 0.6888889 0.7391304 0.7373737
#>  [29] 0.6448598 0.6701031 0.7363636 0.7816092 0.7373737 0.7000000 0.6521739
#>  [36] 0.6842105 0.7604167 0.7826087 0.6853933 0.6170213 0.7522936 0.6236559
#>  [43] 0.6847826 0.6542056 0.7809524 0.7127660 0.7373737 0.6800000 0.6893204
#>  [50] 0.7191011 0.7500000 0.6595745 0.6373626 0.7222222 0.6764706 0.6161616
#>  [57] 0.5900000 0.6372549 0.6734694 0.7142857 0.6826923 0.6576577 0.7093023
#>  [64] 0.6559140 0.7340426 0.7757009 0.6176471 0.6547619 0.6826923 0.7171717
#>  [71] 0.7444444 0.5625000 0.7340426 0.7333333 0.7913043 0.7471264 0.6880734
#>  [78] 0.6702128 0.7009346 0.6956522 0.7383178 0.7128713 0.6804124 0.7075472
#>  [85] 0.7128713 0.7938144 0.7128713 0.6237624 0.6574074 0.7311828 0.6595745
#>  [92] 0.7524752 0.6981132 0.7524752 0.7087379 0.7473684 0.7263158 0.7368421
#>  [99] 0.6909091 0.6634615
#> 
#> $p_treatment_estimate
#>   [1] 0.6857143 0.6354167 0.5769231 0.6666667 0.6534653 0.6336634 0.6422018
#>   [8] 0.5714286 0.5357143 0.6929825 0.7282609 0.7254902 0.5319149 0.6380952
#>  [15] 0.6700000 0.6222222 0.7241379 0.6770833 0.6448598 0.6632653 0.6428571
#>  [22] 0.6222222 0.6132075 0.5961538 0.6568627 0.6909091 0.6759259 0.6237624
#>  [29] 0.6451613 0.6407767 0.7555556 0.5840708 0.7029703 0.7100000 0.6574074
#>  [36] 0.6279070 0.6826923 0.6296296 0.6126126 0.6415094 0.6483516 0.6355140
#>  [43] 0.5833333 0.6881720 0.6105263 0.6603774 0.6138614 0.6100000 0.7319588
#>  [50] 0.6666667 0.6300000 0.6320755 0.6055046 0.6000000 0.6938776 0.6831683
#>  [57] 0.6600000 0.5816327 0.6568627 0.7058824 0.6979167 0.6629213 0.7280702
#>  [64] 0.6635514 0.6886792 0.6559140 0.6428571 0.6120690 0.6145833 0.6336634
#>  [71] 0.6181818 0.7884615 0.6037736 0.6250000 0.6470588 0.6194690 0.7142857
#>  [78] 0.6981132 0.5913978 0.5925926 0.6236559 0.6262626 0.6796117 0.6489362
#>  [85] 0.6464646 0.7087379 0.6666667 0.6464646 0.6847826 0.7383178 0.7169811
#>  [92] 0.6666667 0.6914894 0.7070707 0.6494845 0.6190476 0.7047619 0.5322581
#>  [99] 0.6333333 0.6041667
#> 
#> $prop_diff_estimate
#>   [1] -0.0820132344 -0.0088141026 -0.1498997996  0.0136054422 -0.0726726727
#>   [6] -0.0026005201 -0.0502117362 -0.1749486653 -0.2253462023  0.0185638515
#>  [11]  0.0964343598 -0.1010404162 -0.1325543916 -0.1271794872 -0.0120772947
#>  [16] -0.1177934467  0.0695829095 -0.0400080418 -0.0741708543 -0.0444579781
#>  [21] -0.0052845528 -0.0357869052 -0.0660757454 -0.0390028146 -0.1180944756
#>  [26]  0.0016168149 -0.0676661264 -0.1127758656 -0.0074028378 -0.0286579213
#>  [31]  0.0180423061 -0.2019955203 -0.0364948867  0.0090036014  0.0066451873
#>  [36] -0.0527617477 -0.0733870968 -0.1606344042 -0.0755023341  0.0181818182
#>  [41] -0.1006670710  0.0120603015 -0.1040322581  0.0321866828 -0.1687738310
#>  [46] -0.0502008032 -0.1296032553 -0.0614035088  0.0432779002 -0.0505378526
#>  [51] -0.1208483393 -0.0274989964 -0.0350769231 -0.1205774705  0.0176140913
#>  [56]  0.0660132026  0.0681089744 -0.0541082164 -0.0186858316 -0.0076152305
#>  [61]  0.0194253339  0.0042518728  0.0187985288 -0.0002044572 -0.0424242424
#>  [66] -0.1318116389  0.0216867470 -0.0422997947 -0.0674157303 -0.0862241829
#>  [71] -0.1309186564  0.2263242376 -0.1325301205 -0.1096290121 -0.1460534493
#>  [76] -0.1256363266  0.0262785527  0.0307599517 -0.1114926278 -0.0991935484
#>  [81] -0.1206297281 -0.0842842843 -0.0006006006 -0.0595094491 -0.0694694695
#>  [86] -0.0839511120 -0.0468468468  0.0164824121  0.0298026581  0.0116864800
#>  [91]  0.0593360996 -0.0846944949 -0.0066238458 -0.0472472472 -0.0600600601
#>  [96] -0.1252294514 -0.0285539466 -0.1913080528 -0.0473769825 -0.0625000000
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1]  95 104  96  98  99  99  91 102  88  86 108  98 106  95 100 110 113 104
#>  [19]  93 102  88 110  94  96  98  90  92  99 107  97 110  87  99 100  92 114
#>  [37]  96  92  89  94 109  93  92 107 105  94  99 100 103  89 100  94  91  90
#>  [55] 102  99 100 102  98  98 104 111  86  93  94 107 102  84 104  99  90  96
#>  [73]  94 120 115  87 109  94 107  92 107 101  97 106 101  97 101 101 108  93
#>  [91]  94 101 106 101 103  95  95  76 110 104
#> 
#> $N_treatment
#>   [1] 105  96 104 102 101 101 109  98 112 114  92 102  94 105 100  90  87  96
#>  [19] 107  98 112  90 106 104 102 110 108 101  93 103  90 113 101 100 108  86
#>  [37] 104 108 111 106  91 107 108  93  95 106 101 100  97 111 100 106 109 110
#>  [55]  98 101 100  98 102 102  96  89 114 107 106  93  98 116  96 101 110 104
#>  [73] 106  80  85 113  91 106  93 108  93  99 103  94  99 103  99  99  92 107
#>  [91] 106  99  94  99  97 105 105 124  90  96
#> 

binomialfreq(p_control = 0.5, p_treatment = 0.40, N_total = 200,
            block_number = 2, simulation = 100, drift = -0.15)

#> $power
#> [1] 0
#> 
#> $p_control_estimate
#>   [1] 0.3666667 0.4157303 0.3516484 0.4390244 0.3444444 0.3684211 0.3789474
#>   [8] 0.3617021 0.3409091 0.3689320 0.4020619 0.4339623 0.4022989 0.2947368
#>  [15] 0.4242424 0.3863636 0.3773585 0.3448276 0.3921569 0.4368932 0.3529412
#>  [22] 0.4000000 0.4301075 0.3978495 0.3829787 0.4382022 0.4675325 0.3837209
#>  [29] 0.4777778 0.4186047 0.4333333 0.4090909 0.3186813 0.4157303 0.3617021
#>  [36] 0.4725275 0.3831776 0.3916667 0.3883495 0.2772277 0.3300971 0.3737374
#>  [43] 0.4222222 0.4166667 0.4673913 0.3010753 0.4226804 0.4285714 0.4455446
#>  [50] 0.4117647 0.3548387 0.3750000 0.3444444 0.4421053 0.4158416 0.4130435
#>  [57] 0.4943820 0.3917526 0.4777778 0.4239130 0.4382022 0.3300971 0.3578947
#>  [64] 0.3958333 0.3240741 0.4408602 0.3571429 0.3780488 0.4193548 0.4158416
#>  [71] 0.3235294 0.4408602 0.2777778 0.3584906 0.4516129 0.3200000 0.3296703
#>  [78] 0.3831776 0.4361702 0.4175824 0.3400000 0.3584906 0.3400000 0.3723404
#>  [85] 0.3428571 0.3870968 0.4157303 0.4271845 0.4479167 0.4361702 0.3018868
#>  [92] 0.4285714 0.4356436 0.3838384 0.4183673 0.3608247 0.3831776 0.3958333
#>  [99] 0.5000000 0.4400000
#> 
#> $p_treatment_estimate
#>   [1] 0.2090909 0.2882883 0.3211009 0.2711864 0.3000000 0.2666667 0.2952381
#>   [8] 0.2641509 0.3125000 0.3298969 0.2718447 0.2872340 0.2654867 0.2761905
#>  [15] 0.2376238 0.2410714 0.2553191 0.3274336 0.3163265 0.2474227 0.3877551
#>  [22] 0.3272727 0.2803738 0.3271028 0.2169811 0.2972973 0.1951220 0.3421053
#>  [29] 0.3454545 0.2280702 0.3909091 0.2946429 0.3761468 0.2072072 0.3113208
#>  [36] 0.2660550 0.2688172 0.2625000 0.3195876 0.2828283 0.2371134 0.2574257
#>  [43] 0.3545455 0.2884615 0.2500000 0.3084112 0.3203883 0.3039216 0.2929293
#>  [50] 0.2857143 0.3551402 0.2884615 0.2727273 0.3142857 0.2626263 0.2314815
#>  [57] 0.3063063 0.2038835 0.2636364 0.3055556 0.3153153 0.2886598 0.3428571
#>  [64] 0.2884615 0.2500000 0.3644860 0.2931034 0.3050847 0.2523364 0.1919192
#>  [71] 0.2040816 0.2990654 0.3369565 0.3191489 0.2616822 0.2200000 0.2110092
#>  [78] 0.3010753 0.2358491 0.3302752 0.3400000 0.3191489 0.2600000 0.3207547
#>  [85] 0.3157895 0.3084112 0.3513514 0.2474227 0.3269231 0.2735849 0.3297872
#>  [92] 0.2201835 0.2727273 0.2673267 0.2745098 0.2621359 0.3225806 0.3076923
#>  [99] 0.3240741 0.3100000
#> 
#> $prop_diff_estimate
#>   [1] -0.159660469 -0.134271906 -0.022033556 -0.167837950 -0.038775510
#>   [6] -0.090256410 -0.082213756 -0.104141536 -0.021241830 -0.030778515
#>  [11] -0.134055790 -0.146728222 -0.136317396 -0.017445358 -0.190094245
#>  [16] -0.150326797 -0.119018898  0.002483958 -0.081963928 -0.183991895
#>  [21]  0.034813926 -0.072035613 -0.149138804 -0.070351759 -0.156528792
#>  [26] -0.135173534 -0.263258786 -0.032298137 -0.137250915 -0.181856540
#>  [31] -0.039446930 -0.135620915  0.063427526 -0.201061441 -0.050381373
#>  [36] -0.210002033 -0.112098566 -0.123539232 -0.073864783  0.006401280
#>  [41] -0.086992373 -0.114096651 -0.026762957 -0.128256513 -0.229768198
#>  [46]  0.006633166 -0.100500501 -0.117269076 -0.151959799 -0.102125886
#>  [51]  0.005431503 -0.086538462 -0.071948262 -0.135678392 -0.151155779
#>  [56] -0.190524603 -0.177619532 -0.184131437 -0.195918367 -0.098039216
#>  [61] -0.116992791 -0.054490414 -0.004059265 -0.105410822 -0.066935484
#>  [66] -0.068887518 -0.062012320 -0.076811594 -0.159270517 -0.221097388
#>  [71] -0.112968184 -0.143115578  0.044530297 -0.039357430 -0.185554425
#>  [76] -0.098439376 -0.117362371 -0.084892376 -0.202351987 -0.072667891
#>  [81]  0.000000000 -0.035383997 -0.084455324 -0.053679131 -0.028296207
#>  [86] -0.077372557 -0.065802794 -0.179414733 -0.120993590 -0.158787879
#>  [91]  0.024979855 -0.198367859 -0.160360360 -0.114297173 -0.143714972
#>  [96] -0.098298298 -0.045592705 -0.088141026 -0.174184454 -0.121980676
#> 
#> $N_enrolled
#>   [1] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [19] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [37] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [55] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [73] 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200 200
#>  [91] 200 200 200 200 200 200 200 200 200 200
#> 
#> $N_control
#>   [1]  90  89  91  82  90  95  95  94  88 103  97 106  87  95  99  88 106  87
#>  [19] 102 103 102  90  93  93  94  89  77  86  90  86  90  88  91  89  94  91
#>  [37] 107 120 103 101 103  99  90  96  92  93  97  98 101 102  93  96  90  95
#>  [55] 101  92  89  97  90  92  89 103  95  96 108  93  84  82  93 101 102  93
#>  [73] 108 106  93 100  91 107  94  91 100 106 100  94 105  93  89 103  96  94
#>  [91] 106  91 101  99  98  97 107  96  92 100
#> 
#> $N_treatment
#>   [1] 110 111 109 118 110 105 105 106 112  97 103  94 113 105 101 112  94 113
#>  [19]  98  97  98 110 107 107 106 111 123 114 110 114 110 112 109 111 106 109
#>  [37]  93  80  97  99  97 101 110 104 108 107 103 102  99  98 107 104 110 105
#>  [55]  99 108 111 103 110 108 111  97 105 104  92 107 116 118 107  99  98 107
#>  [73]  92  94 107 100 109  93 106 109 100  94 100 106  95 107 111  97 104 106
#>  [91]  94 109  99 101 102 103  93 104 108 100
#>